==> Stockdale_Hogan_Gilmour_Paddy_Erb/1_QA_All_Streams_2014_2016.Rmd <==
---
title: "All Streams 2014/2016 Genotype QA + FRANz Input with 298 Markers"
author: "Kyle Shedd"
date: "27 January 2022"
output:
  html_notebook:
    theme: united
    toc: yes
editor_options: 
  chunk_output_type: inline
---

```{r setup, include=FALSE}
###############################################################################
#
# HOW TO USE THIS SCRIPT
# This script has been standardized for ease of use with subsequent QA analyses for PINK SALMON
# 1. Copy file to correct stream folder in "V:\Analysis\5_Coastwide\Multispecies\Alaska Hatchery Research Program\PWS Pink\GitHub-PWS-Pink-Parentage" and rename
# 2. Replace names, dates, and other information where applicable in script
#    a. Specific terms requiring replacement in the Introduction and below for future use include: "Stockdale", "PSTOCK", "STOCK", 2017", "2019", "17", "19"
#    b. Please also be sure to update any usernames and passwords, paths, SILLY codes, and filenames, as well as **Results:** text for the notebook
#    c. Do not use this notebook for multiple streams without renaming objects for specificity
#
###############################################################################

rm(list=ls())

knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lubridate)
library(DT)
library(abind)
bbind <- function(...) { abind(..., along = 3) }

.username = readLines("~/../R/usr_pw.txt", n = 1)
.password = readLines("~/../R/usr_pw.txt" , n = 2)[[2]]
source("~/../R/Functions.GCL.R")
```

# Objective

Create a pedigree for all fitness streams for 2014/2016.

# Introduction

This is a quick exploratory analysis of the all 2014/2016 genotype data (Hogan, Stockdale, Gilmour, Paddy, and Erb. Here we will:

  * Read in *LocusControl*
  * Read in all project genotypes
  * Read in paired field/otolith data from [OceanAK](http://www.oceanak.adfg.alaska.gov/)
  * Join genetic data with paired data from OceanAK
    + Will need to create data key for join (DWP barcode + "_" + DWP position number)
  * Determine *FailureRate*
  * Perform QA on project gentoypes
    + Removing fish missing >= 20% of genotypes
    + Remove within-collection duplicates
    + Use a heterozygosity filter to remove individuals with bogus (contaminated) genotype calls
  * Save final, post-QA genotypes with paired data
  * Export genotypes as a *Genepop* file to feed in to *genepop2franz.GCL.r*

**NOTE** this script operates off of the `v1.12.0-static` branch of *GCL-R-Scripts*

# Genotypes

Create *LocusControl* and read in genotypes by *silly*
```{r sillyvec}
(sillyvec <-
   ProjectSillys <-
   paste0(rep(
     c("PHOGAN", "PSTOCK", "PGILMOUR", "PPADDY", "PERB"), each = 2
   ), rep(c("14", "16"), times = 5)))

loci298 <- dget("../../Objects/loci298.txt") # saved list of loci
```

Read in genotypes, one stream at a time, otherwise will never finish

## Hogan

```{r loki_hogan}
read_project_genotypes.GCL(
  sillyvec = sillyvec[1:2],
  loci = loci298,
  username = .username,
  password = .password
)  # LOKI2R works, but we are retaining read_project_genotypes for now as LOKI2R pulls in extraneous info we do not need

if (!dir.exists("../../Genotypes/All_Streams_14_16/")) {
  dir.create("../../Genotypes/All_Streams_14_16/")
}
if (!dir.exists("../../Genotypes/All_Streams_14_16/Raw/")) {
  dir.create("../../Genotypes/All_Streams_14_16/Raw/")
}
save_sillys(sillyvec = sillyvec[1:2], path = "../../Genotypes/Raw/All_Streams_14_16/")
```

## Stockdale

Need to remove `LocusControl`.
```{r loki_stockdale}
rm(LocusControl)

read_project_genotypes.GCL(
  sillyvec = sillyvec[3:4],
  loci = loci298,
  username = .username,
  password = .password
)  # LOKI2R works, but we are retaining read_project_genotypes for now as LOKI2R pulls in extraneous info we do not need

save_sillys(sillyvec = sillyvec[3:4], path = "../../Genotypes/Raw/All_Streams_14_16/")
```

## Gilmour

Need to remove `LocusControl`.
```{r loki_gilmour}
rm(LocusControl)

read_project_genotypes.GCL(
  sillyvec = sillyvec[5:6],
  loci = loci298,
  username = .username,
  password = .password
)  # LOKI2R works, but we are retaining read_project_genotypes for now as LOKI2R pulls in extraneous info we do not need

save_sillys(sillyvec = sillyvec[5:6], path = "../../Genotypes/Raw/All_Streams_14_16/")
```

## Paddy

Need to remove `LocusControl`.
```{r loki_paddy}
rm(LocusControl)

read_project_genotypes.GCL(
  sillyvec = sillyvec[7:8],
  loci = loci298,
  username = .username,
  password = .password
)  # LOKI2R works, but we are retaining read_project_genotypes for now as LOKI2R pulls in extraneous info we do not need

save_sillys(sillyvec = sillyvec[7:8], path = "../../Genotypes/Raw/All_Streams_14_16/")
```

## Erb

Need to remove `LocusControl`.
```{r loki_erb}
rm(LocusControl)

read_project_genotypes.GCL(
  sillyvec = sillyvec[9:10],
  loci = loci298,
  username = .username,
  password = .password
)  # LOKI2R works, but we are retaining read_project_genotypes for now as LOKI2R pulls in extraneous info we do not need

save_sillys(sillyvec = sillyvec[9:10], path = "../../Genotypes/Raw/All_Streams_14_16/")
```

## LocusControl

Create new directories and save objects
```{r save_locus_control_sillys}
loci <- LocusControl$locusnames
nalleles <- LocusControl$nalleles
ploidy <- LocusControl$ploidy
alleles <- LocusControl$alleles

rm(.username, .password, ProjectSillys)

if(!dir.exists("../../Objects/All_Streams_14_16/")) {
  dir.create("../../Objects/All_Streams_14_16/")
}

save_objects(objects = c("LocusControl", "loci", "sillyvec"),
             path = "../../Objects/All_Streams_14_16/")
```

# OceanAK

## Read and format

Read in paired field and otolith data from *OceanAK* and join that data to the attributes table of each silly object
```{r read_oceanak, message=FALSE}
# The `genepop2franz.GCL` function we use in the very last step expects the *OceanAK* file to have the "old" names (lowercase with spaces) and not the "new" names (UPPERCASE with _); we convert the column names to the old format here. Note that this is a patch and column names should be double-checked during each new analysis.
og_names <-
  suppressMessages(names(
    read_csv(file = "../../OceanAK/PedigreeData_AHRP - Salmon Biological Data 2_PWS_2013-2018_no_otoliths.csv", progress = FALSE)
  )) # create object with the old names; do not update path - it is simply a file from which to pull the old names
oceanak <-
  read_csv(file = "../../OceanAK/AHRP Salmon Biological Data 20220124_151655.csv") # read in OceanAK data; recommend new data pull prior to running
```

Check names to confirm if they match up
```{r oceanak_name_check}
list(og = og_names, new = names(oceanak))
```

Whoops, forgot that we have some new columns associated with the GOD event, gonna drop those from the re-name
```{r oceanak_name_compare}
tibble(OG = og_names, new = names(oceanak)[seq(length(og_names))])
```

Rename, drop GOD, save
```{r oceanak_rename}
names(oceanak)[seq(length(og_names))] <- og_names # rename columns

oceanak <- oceanak %>% select(-contains("TARGET"), -contains("DETERMINATION"), -contains("ARRAY"))

write_csv(oceanak, "../../OceanAK/AHRP Salmon Biological Data 20220124_151655_names.csv") # output renamed OceanAK data
```

## Visualize and subset

```{r}
oceanak %>%
  count(`Sample Year`, `Location Code`) %>%
  pivot_wider(names_from = `Sample Year`,
              values_from = n,
              values_fill =  0)
```

Subset the data by streams and years of interest
```{r}
(
  all_streams_14_16_oceanak <- oceanak %>%
    dplyr::filter(`Sample Year` %in% c(2014, 2016)) %>%
    tidyr::unite(
      silly_source,
      c("Silly Code", "Fish ID"),
      sep = "_",
      remove = FALSE
    )
)
```

Count otolith marks present/absent/NA by year
```{r}
all_streams_14_16_oceanak %>% 
  count(`Location Code`, `Sample Year`, `Otolith Mark Present`) %>% 
  pivot_wider(names_from = `Otolith Mark Present`,
              values_from = n,
              values_fill =  0)
```

Verify that all genotypes are in OceanAK data
```{r}
sapply(sillyvec, function(silly) {
  my.gcl = get(paste0(silly, ".gcl"))
  table(my.gcl$attributes$SillySource %in% all_streams_14_16_oceanak$silly_source)
}, simplify = FALSE)
```

Fabulous, all in there. That is the huge upside of working with the data warehouse that Tim and Eric put together.

List attribute names, since we used `read_project_genotypes.GCL` instead of `LOKI2R.GCL`?
```{r}
names(PSTOCK14.gcl$attributes)
```

## Join attributes table

Join *all_streams_14_16_oceanak* data to the attributes table of each silly. Already looked to verify that no tissues were missing or duplicated, which makes sense because the lab can't genotype tissue that aren't there and they never genotype two samples in a well.
```{r join_oceanak}
sapply(sillyvec, function(silly) {
  my.gcl = get(paste0(silly, ".gcl"))
  
  my.gcl$attributes <- my.gcl$attributes %>%
    dplyr::left_join(all_streams_14_16_oceanak,
                     by = c("SillySource" = "silly_source")) %>%
    dplyr::select(
      FK_FISH_ID,
      SILLY_CODE,
      PLATE_ID,
      `DNA Tray Code`,
      `DNA Tray Well Code`,
      SillySource,
      `Sample Date`,
      Sex,
      `Length Mm`,
      `Location Code`,
      `Otolith Mark Present`,
      `Otolith Mark ID`,
      `Sample Year`
    ) %>%
    dplyr::rename(DNA_TRAY_CODE = "DNA Tray Code",
                  DNA_TRAY_WELL_CODE = "DNA Tray Well Code")
  
  assign(x = paste0(silly, ".gcl"),
         value = my.gcl,
         pos = 1)
}, simplify = TRUE)
```

Save
```{r}
if(!dir.exists("../../Genotypes/All_Streams_14_16/OceanAK")) {
  dir.create("../../Genotypes/All_Streams_14_16/OceanAK")
}

save_sillys(sillyvec = sillyvec, path = "../../Genotypes/All_Streams_14_16/OceanAK")
```

## Split by otolith-origin

First, we need to verify that we only have hatchery and natural-origin fish for the parental year (2014), and remove any known hatchery-origin fish from the offspring year (2016). Normally we'd just keep natural-origin fish from the offspring year (2016), but so many of these samples were involved in the GOD event, the have NA.

### 2014 - Parents

Split into hatchery vs. natural, NAs are dropped automagically
```{r}
sapply(grep(pattern = "14", x = sillyvec, value = TRUE), function(silly) {
  my.gcl = get(paste0(silly, ".gcl"))
  
  # get natural_ids
  natural_ids <- list(
    my.gcl$attributes %>%
      filter(`Otolith Mark Present` == "NO") %>%
      mutate(FK_FISH_ID = as.character(FK_FISH_ID)) %>%
      pull(FK_FISH_ID)
  )
  names(natural_ids) = silly
  
  # pool natural
  PoolCollections.GCL(
    collections = silly,
    loci = loci,
    IDs = natural_ids,
    newname = paste0(silly, "n")
  )
  
  # get hatchery_ids
  hatchery_ids <- list(
    my.gcl$attributes %>%
      filter(`Otolith Mark Present` == "YES") %>%
      mutate(FK_FISH_ID = as.character(FK_FISH_ID)) %>%
      pull(FK_FISH_ID)
  )
  names(hatchery_ids) = silly
  
  # pool hatchery
  PoolCollections.GCL(
    collections = silly,
    loci = loci,
    IDs = hatchery_ids,
    newname = paste0(silly, "h")
  )
  
})
```

### 2016 - Offspring

Just drop known hatchery-origin fish (so many NAs due to GOD event)
```{r}
sapply(grep(pattern = "16", x = sillyvec, value = TRUE), function(silly) {
  my.gcl = get(paste0(silly, ".gcl"))
  
  # get hatchery_ids
  hatchery_ids <- list(
    my.gcl$attributes %>%
      filter(`Otolith Mark Present` == "YES") %>%
      mutate(FK_FISH_ID = as.character(FK_FISH_ID)) %>%
      pull(FK_FISH_ID)
  )
  names(hatchery_ids) = silly
  
  # remove hatchery
  RemoveIDs.GCL(silly = silly, IDs = hatchery_ids)
  
})
```

Create new *sillyvec*
```{r}
(sillyvec_parentage <-
   paste0(rep(
     c("PHOGAN", "PSTOCK", "PGILMOUR", "PPADDY", "PERB"), each = 3
   ), rep(c("14n", "14h", "16"), times = 5)))

save_objects(objects = "sillyvec_parentage", path = "../../Objects/All_Streams_14_16/")
```

Verify that we did as intended, check sillys for otolith read counts
```{r}
sapply(sillyvec_parentage, function(silly) {
  my.gcl = get(paste0(silly, ".gcl"))
  my.gcl$attributes %>%
    select(`Otolith Mark Present`) %>% 
    mutate(silly = silly)
}, simplify = FALSE) %>%
  dplyr::bind_rows() %>%
  count(silly, `Otolith Mark Present`) %>%
  pivot_wider(names_from = `Otolith Mark Present`,
              values_from = n,
              values_fill =  0)
```

Save
```{r}
if(!dir.exists("../../Genotypes/All_Streams_14_16/OceanAK_Origin")) {
  dir.create("../../Genotypes/All_Streams_14_16/OceanAK_Origin")
}

save_sillys(sillyvec = sillyvec_parentage, path = "../../Genotypes/All_Streams_14_16/OceanAK_Origin")
```

# Failure rate

Calculate and view the failure rate by silly, locus, and plate
```{r failure_rate}
project <- "all_streams_14_16" #this is required of the function; can be any name
failure_rate <- FailureRate.GCL(sillyvec = sillyvec_parentage)
failure_rate_noplots <- failure_rate[1:4]
save_objects("failure_rate_noplots", "../../Objects/All_Streams_14_16")
failure_rate_noplots
```
**Results:** There's an overall failure rate of **8%**. There are some specific plate effects, but they are very scattered, as are any locus-specific failures. 

# QA

Go through our standard Quality Assurance (QA) process, explore heterozygosity, and produce a matrix with sample sizes.
```{r qa_setup}
ProjectSillys_Origin <- sillyvec_parentage

ProjectSillys_Origin_SampleSizes <-
  matrix(
    data = NA,
    nrow = length(ProjectSillys_Origin),
    ncol = 5,
    dimnames = list(
      ProjectSillys_Origin,
      c("Genotyped", "Missing", "Duplicate", "Heterozygosity", "Final")
    )
  )

ProjectSillys_Origin_SampleSizes[, "Genotyped"] <-
  sapply(paste0(ProjectSillys_Origin, ".gcl"), function(x)
    get(x)$n)
```

## Filter missing

Remove individuals missing >=20% of genotypes (i.e. the 80% rule)
```{r qa_missing}
MissLoci <-
  RemoveIndMissLoci.GCL(sillyvec = ProjectSillys_Origin, proportion = 0.8)

# MissLoci  # character vector of fish removed

save_objects("MissLoci", path = "../../Objects/All_Streams_14_16/")

ProjectSillys_Origin_SampleSizes[, "Missing"] <-
  ProjectSillys_Origin_SampleSizes[, "Genotyped"] - sapply(paste0(ProjectSillys_Origin, ".gcl"), function(x)
    get(x)$n)
```

## Filter duplicates

Remove duplicate individuals within the same collection. Typically, we specify *duplicates* as a pair of individuals that share >=95% of genotypes. Once a pair of *duplicates* is identified, we keep the individual with the most genotypes and remove the other(s).

Identify duplicates
```{r qa_duplicate}
DuplicateCheck95MinProportion <-
  CheckDupWithinSilly.GCL(
    sillyvec = ProjectSillys_Origin,
    loci = loci,
    quantile = NULL,
    minproportion = 0.95
  )

DuplicateCheckReportSummary <-
  sapply(ProjectSillys_Origin, function(x)
    DuplicateCheck95MinProportion[[x]]$report, simplify = FALSE)

DuplicateCheckReportSummary

save_objects("DuplicateCheckReportSummary", path = "../../Objects/All_Streams_14_16/")
```

Remove **both** duplicates! As opposed to GSI work, where we want to keep individuals but aren't typically worried about paried data, here we want to remove both individuals as the paired data integrity is lost.

Grab duplicate fish_ids from `ID1` and `ID2`
```{r gather_duplicates}
#RemovedDups <- RemoveDups.GCL(DuplicateCheck95MinProportion)
#save_objects("RemovedDups", path = "../Objects/")

# All duplicate IDs in to one tibble
(
  dups <-
    bind_rows(DuplicateCheckReportSummary[sapply(DuplicateCheckReportSummary, is.data.frame)], .id = "silly") %>%
    select(silly, ID1, ID2) %>%
    pivot_longer(
      names_to = "trash",
      values_to = "id",
      cols = -silly
    ) %>%
    select(-trash) %>%
    arrange(silly, id)
)
```

Separate out by silly
```{r split_duplicates}
# One tibble with all attributes for fish to remove
(
  duplicates_to_remove <-
    sapply(unique(dups$silly), function(mysilly) {
      dup_ids <- filter(dups, silly == mysilly) %>% pull(id)
      
      my.gcl <- get(paste0(mysilly, ".gcl"))
      
      dup_attributes <- my.gcl$attributes %>%
        filter(FK_FISH_ID %in% dup_ids) %>%
        mutate(FK_FISH_ID = as.numeric(FK_FISH_ID))
      
      dup_attributes
    }, simplify = FALSE) %>%
    dplyr::bind_rows() %>%
    as_tibble()
)

save_objects("duplicates_to_remove", path = "../../Objects/All_Streams_14_16/")
```

Remove duplicates as identified
```{r remove_duplicates}
sapply(unique(dups$silly), function(mysilly) {
  
  dup_ids <- filter(dups, silly == mysilly) %>% pull(id)
  
  RemoveIDs.GCL(silly = mysilly, IDs = dup_ids)
})

# Fill in sample size matrix
ProjectSillys_Origin_SampleSizes[, "Duplicate"] <-  
  ProjectSillys_Origin_SampleSizes[, "Genotyped"] - 
  ProjectSillys_Origin_SampleSizes[, "Missing"] - 
  sapply(paste0(ProjectSillys_Origin, ".gcl"), function(x) get(x)$n)
```

## Filter for heterozygosity

We know from the QC work that the **GTscore** genotyping pipeline can sometimes result in individuals with an overabundance of heterozygous calls (perhaps from contamination). We want to remove individuals with untrustworthy genotypes. To explore the utility of a *heterozygosity filter*, we should first plot a histogram of heterozygosity per individual to see what the problem looks like. From there we can set a threshold or range of acceptable heterozygosities.

Calculate heterozygosity

**NOTE** the *old* way was overly simplistic and ignored no-calls (i.e. score = "0")
```{r heterozygosity_calculations}
(
  het_tibble <- sapply(ProjectSillys_Origin, function(mysilly) {
    my.gcl <- get(paste0(mysilly, ".gcl"))
    
    ind_het <-
      apply(my.gcl$scores, 1, function(ind) {
        sum(ind[, "Dose1"] != ind[, "Dose2"]) / sum(ind[, "Dose1"] != "0")
      })
    
    ind_het_old <-
      apply(my.gcl$scores, 1, function(ind) {
        sum(ind[, "Dose1"] != ind[, "Dose2"]) / 298
      })
    
    tibble(
      silly = mysilly,
      og_silly = my.gcl$attributes$SILLY_CODE,
      fish_id = as.character(my.gcl$attributes$FK_FISH_ID),
      sillysource = my.gcl$attributes$SillySource,
      het = ind_het,
      old_het = ind_het_old
    )
  }, simplify = FALSE) %>%
    dplyr::bind_rows() %>% 
    mutate(lineage = "even")
)

save_objects(objects = "het_tibble", path = "../../Objects/All_Streams_14_16/")
```

### Plot heterozygosities

Plot heterozygosities by lineage with *tidyverse* for ease of use/visualization. Cutoffs (red vertical lines) are based on the +/- 1.5 IQR method that it is well known and well documented in literature.

#### Compare method

Compare distribution of heterozygosities calculated by the *new* method vs. the *old*, naive method
**NOTE** no need to incorporate lineage here, since all collections here are even-year
```{r}
# NOTE: Original plotting by silly removed and +/- 1.5 IQR heterozygosity cutoff instilled -- After discussions among Chris, Tyler, Dan, and Kyle, they decided it would be more appropriate to examine heterozygosity by lineage (even and/or odd) instead of year. It was also agreed that the +/- 3SD cutoff (Anderson et al. 2010) was not restrictive enough as the distribution has a long right-sided tail and the upper bound for the cutoff should be closer to the main distribution. A hybrid right-side mirror approach based on the +/- 3SD was also suggested by Chris and discarded in favor of the +/- 1.5 IQR.

het_method_means <- het_tibble %>%
  pivot_longer(names_to = "het_method",
               values_to = "het_ind",
               cols = contains("het")) %>%
  group_by(het_method) %>%
  summarise(
    het_mean = mean(het_ind),
    het_1.5IQR_low = quantile(het_ind, probs = 0.25) - 1.5 * diff(quantile(het_ind, probs = c(0.25, 0.75))),
    het_1.5IQR_high = quantile(het_ind, probs = 0.75) + 1.5 * diff(quantile(het_ind, probs = c(0.25, 0.75)))
  ) %>% 
  mutate(lineage = "even")

het_tibble %>%
  tidyr::pivot_longer(names_to = "het_method",
                      values_to = "het_ind",
                      cols = contains("het")) %>%
  ggplot2::ggplot(aes(x = het_ind)) +
  ggplot2::geom_histogram(binwidth = 1 / length(loci)) +
  ggplot2::facet_grid(rows = vars(het_method)) +
  ggplot2::geom_vline(data = het_method_means, aes(xintercept = het_mean, colour = "mean")) +
  ggplot2::geom_vline(data = het_method_means, aes(xintercept = het_1.5IQR_low, colour = "IQR")) +
  ggplot2::geom_vline(data = het_method_means, aes(xintercept = het_1.5IQR_high, colour = "IQR")) +
  ggplot2::scale_color_manual(name = "statistics", values = c(mean = "blue", IQR = "red")) +
  ggplot2::xlim(0, 1) +
  ggplot2::xlab("Individual Heterozygosity") +
  ggplot2::ylab("Frequency") +
  ggplot2::ggtitle("Individual Heterozygosity By Method") +
  ggplot2::theme_bw() 
```

#### Compare silly

See how the *new* method would apply across original sillys
```{r fig.height=10}
# het_silly_means <- het_tibble %>%
#   group_by(og_silly) %>%
#   summarise(
#     het_mean = mean(het),
#     het_1.5IQR_low = quantile(het, probs = 0.25) - 1.5 * diff(quantile(het, probs = c(0.25, 0.75))),
#     het_1.5IQR_high = quantile(het, probs = 0.75) + 1.5 * diff(quantile(het, probs = c(0.25, 0.75)))
#   )

het_tibble %>%
  ggplot2::ggplot(aes(x = het)) +
  ggplot2::geom_histogram(binwidth = 1 / length(loci)) +
  ggplot2::facet_grid(rows = vars(og_silly), scales = "free_y") +
  ggplot2::geom_vline(
    data = filter(het_method_means, het_method == "het"),
    aes(xintercept = het_mean, colour = "mean")
    ) +
  ggplot2::geom_vline(
    data = filter(het_method_means, het_method == "het"),
    aes(xintercept = het_1.5IQR_low, colour = "IQR")
  ) +
  ggplot2::geom_vline(
    data = filter(het_method_means, het_method == "het"),
    aes(xintercept = het_1.5IQR_high, colour = "IQR")
  ) +
  ggplot2::scale_color_manual(name = "statistics", values = c(mean = "blue", IQR = "red")) +
  ggplot2::xlim(0, 1) +
  ggplot2::xlab("Individual Heterozygosity") +
  ggplot2::ylab("Frequency") +
  ggplot2::ggtitle("Individual Heterozygosity By Method") +
  ggplot2::theme_bw(base_size = 12) 
```

### Remove fish outside of the IQR cutoffs

Count the number of fish to remove from each silly
```{r table_het_cutoff}
# Calculate 1.5 IQR
het_method_means

# Get list of fish to remove
het_tibble %>%
  dplyr::left_join(filter(het_method_means, het_method == "het"), by = "lineage") %>%
  dplyr::filter(het < het_1.5IQR_low | het > het_1.5IQR_high) %>%
  dplyr::count(og_silly)
```

Filter for fish to remove and save
```{r remove_het_cutoff}
(
  removed_hets <- het_tibble %>%
    dplyr::left_join(filter(het_method_means, het_method == "het"), by = "lineage") %>%
    dplyr::filter(het < het_1.5IQR_low | het > het_1.5IQR_high)
)

save_objects("removed_hets", "../../Objects/All_Streams_14_16/")
```

Remove these fish. Rather than convert *SillySource* to *FK_FISH_ID* (**note**: the values for *FK_FISH_ID* got reassigned after pooling) before removal, I retained the silly codes and fish_ids.
```{r}
sapply(unique(removed_hets$silly), function(mysilly) {
  
  het_ids <- filter(removed_hets, silly == mysilly) %>% pull(fish_id)
  
  RemoveIDs.GCL(silly = mysilly, IDs = het_ids)
})

ProjectSillys_Origin_SampleSizes[, "Heterozygosity"] <-
  ProjectSillys_Origin_SampleSizes[, "Genotyped"] - 
  ProjectSillys_Origin_SampleSizes[, "Missing"] - 
  ProjectSillys_Origin_SampleSizes[, "Duplicate"] - 
  sapply(paste0(ProjectSillys_Origin, ".gcl"), function(x) get(x)$n)
```

## Final dataset

Count how many fish we retained
```{r qa_final}
ProjectSillys_Origin_SampleSizes[, "Final"] <-
  sapply(paste0(ProjectSillys_Origin, ".gcl"), function(x) get(x)$n)

ProjectSillys_Origin_SampleSizes

save_objects("ProjectSillys_Origin_SampleSizes", path = "../../Objects/All_Streams_14_16/")
```

Save the post-QA genotypes
```{r save_postQA_genotypes}
if(!dir.exists("../../Genotypes/All_Streams_14_16/OceanAK_Origin_PostQA")) {
  dir.create("../../Genotypes/All_Streams_14_16/OceanAK_Origin_PostQA")
}

save_sillys(ProjectSillys_Origin, "../../Genotypes/All_Streams_14_16/OceanAK_Origin_PostQA/")
```

# Export genepop file

Now that we have final post QA/QC genotypes, we want to save our work by exporting the genotypes as a **Genepeop** file using `gcl2Genepop.GCL`, with individuals named by *SillySource*.
```{r genepop}
if(!dir.exists("../../Genepop/All_Streams_14_16")) {dir.create("../../Genepop/All_Streams_14_16")}

gcl2Genepop.GCL(sillyvec = ProjectSillys_Origin, 
                loci = loci, 
                path = "../../Genepop/All_Streams_14_16/All_Streams_14_16_postQA.gen", 
                VialNums = TRUE, 
                usat = FALSE)
```

# Convert genepop to FRANz format

The **Genepop** file is fed into Chase's *genepop2franz.GCL.r* script to create a **FRANz** input file for parentage analysis.
```{r Convert genepop to Franz for parentage}
# detach(package:reshape2, unload = TRUE)
source("../genepop2franz.GCL.r")
genepop2franz.GCL(
  Genepop = "../../Genepop/All_Streams_14_16/All_Streams_14_16_postQA.gen",
  OceanAK = "../../OceanAK/AHRP Salmon Biological Data 20220124_151655_names.csv",
  Year = c(2014, 2016),
  Stream = c("HOGAN", "STOCK", "GILMOUR", "PADDY", "ERB"),
  output_dir = "../"
)
```

**NOTE** Since we want to run all 5 streams together, you need to go in and edit the *FRANz* input file to reflect 1 sampling location and take out delimiters between locations.

Final sample size
```{r}
sum(ProjectSillys_Origin_SampleSizes[, "Final"])
```

End
