==> Stockdale_Hogan_Gilmour_Paddy_Erb/1_QA_All_Streams_2016_2018_develop.Rmd <==
---
title: "All Streams 2016/2018 Genotype QA + FRANz Input with 298 Markers"
subtitle: "`develop` branch GCL-R-Scripts"
author: "Kyle Shedd"
date: "12 May 2022"
output:
  html_notebook:
    theme: united
    toc: yes
editor_options: 
  chunk_output_type: inline
---

```{r setup, include=FALSE}
###############################################################################
#
# HOW TO USE THIS SCRIPT
# This script has been standardized for ease of use with subsequent QA analyses for PINK SALMON
# This script assumes just two years (parent/offspring), but could be adapted for analysis of multiple generations/lineages
# This script has been re-written and optimized for the "new" GCL-R-Scripts (tibble vs. list) from the `develop` or `master` branch!
# 1. Copy file to correct stream folder in "V:\Analysis\5_Coastwide\Multispecies\Alaska Hatchery Research Program\PWS Pink\GitHub-PWS-Pink-Parentage" and rename
# 2. Replace names, dates, and other information where applicable in script - it is modeled off of "1_QA_All_Streams_2014_2016_develop.Rmd"
#    a. Specific terms requiring replacement in the Introduction and below for future use include: 
#      * "all_streams_14_16" (name of oceanak)
#      * "All_Streams_14_16" (name of "Objects" and "Gentoypes" subdirectories)
#      * sillyvec: c("PHOGAN", "PSTOCK", "PGILMOUR", "PPADDY", "PERB")
#      * Stream (genepop2franz.GCL): c("HOGAN", "STOCK", "GILMOUR", "PADDY", "ERB")
#      * years: "2014", "2016", "14", "16"
#    b. Please also be sure to update any usernames and passwords, paths, SILLY codes, and filenames (e.g. OceanAK), as well as **Results:** text for the notebook
#    c. This notebook is set up to handle multiple streams, but not currently multiple generations (just F0/F1 for now).
#
###############################################################################

rm(list=ls())

knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lubridate)
library(DT)

.username = readLines("~/../R/usr_pw.txt", n = 1)
.password = readLines("~/../R/usr_pw.txt" , n = 2)[[2]]
source("~/../R/Functions.GCL.R")  # develop branch!!!
```

# Objective

Create a pedigree for all fitness streams for 2016/2018.

# Introduction

This is a quick exploratory analysis of the 2016/2018 genotype data from all five fitness streams (Hogan, Stockdale, Gilmour, Paddy, and Erb). Here we will:

  * Read in *LocusControl*
  * Read in all project genotypes
  * Read in paired field/otolith data from [OceanAK](http://www.oceanak.adfg.alaska.gov/)
  * Join genetic data with paired data from OceanAK
    + Will need to create data key for join (SillySource, Silly_FishID)
  * Perform QA on project genotypes (original SILLY_CODE)
    + Removing fish missing >= 20% of genotypes
    + Remove within-collection duplicates at 95% (both)
    + Use a heterozygosity filter to remove individuals with bogus (contaminated) genotype calls 1.5 IQR
  * Save final, post-QA genotypes with paired data
  * Export genotypes as a *Genepop* file to feed in to *genepop2franz.GCL.r*

**NOTE** this script operates off of the `develop` branch of *GCL-R-Scripts*

# Genotypes

Create *LocusControl* and read in genotypes by *silly*
```{r sillyvec}
(sillyvec <-
     paste0(rep(
     c("PHOGAN", "PSTOCK", "PGILMOUR", "PPADDY", "PERB"), each = 2
   ), rep(c("16", "18"), times = 5)))

loci298 <- dget("../../Objects/loci298.txt") # saved list of loci
```

## LocusControl

Create `LocusControl`, new Objects directory, and save objects
```{r LocusControl}
CreateLocusControl.GCL(locusnames = loci298,
                       username = .username,
                       password = .password)

loci <- LocusControl$locusnames

if(!dir.exists("../../Objects/All_Streams_16_18_develop/")) {
  dir.create("../../Objects/All_Streams_16_18_develop/")
}

save_objects(objects = c("LocusControl", "loci", "sillyvec"),
             path = "../../Objects/All_Streams_16_18_develop/")
```

## Genotypes

Read in genotypes
```{r}
LOKI2R.GCL(sillyvec = sillyvec,
           username = .username,
           password = .password)
```

Save genotypes
```{r}
if (!dir.exists("../../Genotypes/All_Streams_16_18_develop/")) {
  dir.create("../../Genotypes/All_Streams_16_18_develop/")
}

if (!dir.exists("../../Genotypes/All_Streams_16_18_develop/Raw/")) {
  dir.create("../../Genotypes/All_Streams_16_18_develop/Raw/")
}

save_sillys(sillyvec = sillyvec, path = "../../Genotypes/All_Streams_16_18_develop/Raw/")
```

Load genotypes (if re-running)
```{r}
load_sillys(path = "../../Genotypes/All_Streams_16_18_develop/Raw/")
```

# OceanAK

## Read and format

Read in paired field and otolith data from *OceanAK* and join that data to the attributes table of each silly object
```{r read_oceanak, message=FALSE}
# The `genepop2franz.GCL` function we use in the very last step expects the *OceanAK* file to have the "old" names (lowercase with spaces) and not the "new" names (UPPERCASE with _); we convert the column names to the old format here. Note that this is a patch and column names should be double-checked during each new analysis.
og_names <-
  suppressMessages(names(
    readr::read_csv(file = "../../OceanAK/PedigreeData_AHRP - Salmon Biological Data 2_PWS_2013-2018_no_otoliths.csv", progress = FALSE)
  )) # create object with the old names; do not update path - it is simply a file from which to pull the old names
oceanak <-
  readr::read_csv(file = "../../OceanAK/AHRP Salmon Biological Data 20220511_174609.csv") # read in OceanAK data; recommend new data pull prior to running
```

### Duplicates?

Before renameing those columns, check to see if we have any duplicates in `oceanak`
```{r}
(
  oceanak_duplicates <- oceanak %>%
    tidyr::unite(
      silly_source,
      c("SILLY_CODE", "FISH_ID"),
      sep = "_",
      remove = FALSE
    ) %>%
    dplyr::group_by(silly_source) %>%
    dplyr::filter(dplyr::n() > 1)
)

oceanak_duplicates %>% 
  dplyr::distinct(silly_source, SILLY_CODE, DNA_TRAY_CODE) %>% 
  dplyr::ungroup() %>% 
  dplyr::count(SILLY_CODE, DNA_TRAY_CODE)
```

Whew, all resolved :)

```{r}
# oceanak_duplicates %>% 
#   dplyr::select(silly_source, SAMPLE_ID, dplyr::contains("OTOLITH"))
```

Well, shit. We do. I'm going to write out a .csv to send to Tim Frawley and figure out WTF is going on here...
```{r}
# readr::write_csv(oceanak_duplicates, "../../OceanAK/AHRP Salmon Biological Data 20220426_092121_duplicates.csv")
```

#### 0000022639

From Tim, it looks like SAMPLE_ID 201799494 is DWP 0000022639 and SAMPLE_ID 201799955 should be 0000022629.

Investigate whether DWP 0000022629 has any otolith reads
```{r}
# oceanak %>% 
#   dplyr::filter(DNA_TRAY_CODE == "0000022629") %>% 
#   dplyr::select(SILLY_CODE, DNA_TRAY_CODE, SAMPLE_ID, dplyr::contains("OTOLITH"))
```

Nope, no data from this PGILMOUR17 DWP. I double checked the export data (11/8/2017) and it lines up with the read date (5/4/2018)

Sample ID: 201799955 , (otolith) Tray Number: 0000022629, current DNA Tray Code: 0000022639 should be changed to DNA Tray Code: 0000022629.

#### 0000028424

From Tim, it looks like TRAY_NUMBER 0000028401 is coded as DWP 0000028424 as well as TRAY_NUMBER 0000028424 is 0000028424

Investigate whether DWP 0000028401 has any otolith reads
```{r}
# oceanak %>% 
#   dplyr::filter(DNA_TRAY_CODE == "0000028401") %>% 
#   dplyr::select(DNA_TRAY_CODE, SAMPLE_ID, dplyr::contains("OTOLITH"))
```

Nope, same issue as above.

Sample ID: 201800235 , (otolith) Tray Number: 0000028401, current DNA Tray Code: 0000028424 should be changed to DNA Tray Code: 0000028401.

**RESOLUTION**
Worked with Tim Frawley, Jenni Morella, and Megan Lovejoy to resolve this issue on 4/26/22. Update should take place at end of April.

### Re-name columns

Check names to confirm if they match up
```{r oceanak_name_check}
list(og = og_names, new = names(oceanak))
```

Whoops, forgot that we have some new columns associated with the GOD event, gonna drop those from the re-name
```{r oceanak_name_compare}
tibble(OG = og_names, new = names(oceanak)[seq(length(og_names))])
```

Rename, drop GOD columns, save (will need this later to create FRANz input file)
```{r oceanak_rename}
names(oceanak)[seq(length(og_names))] <- og_names # rename columns

oceanak <-
  oceanak %>% select(-contains("TARGET"),
                     -contains("DETERMINATION"),
                     -contains("ARRAY"))

write_csv(oceanak, "../../OceanAK/AHRP Salmon Biological Data 20220511_174609_names.csv") # output renamed OceanAK data
```

## Visualize and subset

How many samples by stream and year?
```{r}
oceanak %>%
  count(`Sample Year`, `Location Code`) %>%
  pivot_wider(names_from = `Sample Year`,
              values_from = n,
              values_fill =  0)
```

Subset the data by streams and years of interest
```{r}
(
  all_streams_16_18_oceanak <- oceanak %>%
    dplyr::filter(`Sample Year` %in% c(2016, 2018)) %>%
    tidyr::unite(
      silly_source,
      c("Silly Code", "Fish ID"),
      sep = "_",
      remove = FALSE
    )
)
```

Do we have any duplicates in `all_streams_16_18_oceanak`?
```{r}
table(table(all_streams_16_18_oceanak$silly_source))
```

Nope, not anymore! (previously there were 48)

Count otolith marks present/absent/NA by year
```{r}
all_streams_16_18_oceanak %>% 
  count(`Location Code`, `Sample Year`, `Otolith Mark Present`) %>% 
  pivot_wider(names_from = `Otolith Mark Present`,
              values_from = n,
              values_fill =  0)
```

Verify that all genotypes are in OceanAK data
```{r}
sapply(sillyvec, function(silly) {
  my.gcl = get(paste0(silly, ".gcl"))
  table(my.gcl$SillySource %in% all_streams_16_18_oceanak$silly_source)
}, simplify = FALSE)
```

Fabulous, all in there. That is the huge upside of working with the data warehouse that Tim and Eric put together.

How about the other way around? Which samples from these streams/years have NOT been genotyped?
```{r}
lapply(objects(pattern = "\\.gcl"), get) %>%
  bind_rows() %>%
  select(SillySource) %>%
  anti_join(
    x = all_streams_16_18_oceanak,
    y = .,
    by = c("silly_source" = "SillySource")
  ) %>%
  count(`Location Code`, `Sample Year`, `Otolith Mark Present`) %>%
  pivot_wider(names_from = `Otolith Mark Present`,
              values_from = n,
              values_fill =  0)
```

**Note** that we did not genotype all hatchery-origin samples available in 2018 from  Hogan. This decision was made at the time due to the unbalanced hatchery/natural sample sizes and the strict samples sizes budgeted for our external grants (NPRB and S/K). Otherwise, there are a smattering of samples that got passed over (missing sex data or flagged as tissue missing), likely due to NA origin reads at the time that later got 2nd read.

Remove `oceanak` for memory
```{r}
rm(oceanak)
```

## Join attributes table

Join *all_streams_16_18_oceanak* data to the attributes table of each silly. Already looked to verify that no tissues were missing or duplicated, which makes sense because the lab can't genotype tissue that aren't there and they never genotype two samples in a well.
```{r join_oceanak}
invisible(sapply(sillyvec, function(silly) {
  my.gcl = get(paste0(silly, ".gcl"))
  
  my_new.gcl <- my.gcl %>%
    dplyr::left_join(
      dplyr::select(  # just grab what we need from OceanAK data, not everything!
        .data = all_streams_16_18_oceanak,
        silly_source,
        `Location Code`,
        `Sample Year`,
        `Sample Date`,
        Sex,
        `Length Mm`,
        `Otolith Mark Present`,
        `Otolith Mark ID`
      ),
      by = c("SillySource" = "silly_source")
    )
  
  assign(x = paste0(silly, ".gcl"),
         value = my_new.gcl,
         pos = 1)
}, simplify = FALSE))
```

Save
```{r}
if(!dir.exists("../../Genotypes/All_Streams_16_18_develop/OceanAK")) {
  dir.create("../../Genotypes/All_Streams_16_18_develop/OceanAK")
}

save_sillys(sillyvec = sillyvec, path = "../../Genotypes/All_Streams_16_18_develop/OceanAK")
```

Remove `all_streams_16_18_oceanak` for memory
```{r}
rm(all_streams_16_18_oceanak)
```

## Remove fish based on otolith data

First, we need to verify that we only have hatchery and natural-origin fish for the parental year (2016), and remove any known hatchery-origin fish from the offspring year (2018). Normally we'd just keep natural-origin fish from the offspring year (2018), but so many of these samples were involved in the GOD event, the have NA.

### 2016 - Parents

Just drop NAs from parent generation, we only want to keep known natural- and hatchery-origin fish
```{r}
invisible(sapply(grep(pattern = "16", x = sillyvec, value = TRUE), function(silly) {
  my.gcl = get(paste0(silly, ".gcl"))
  
  # drop hatchery, keep natural and NA
  assign(
    x = paste0(silly, ".gcl"),
    value = filter(my.gcl, !is.na(`Otolith Mark Present`)),
    pos = 1
  )
  
}))
```

### 2018 - Offspring

Just drop known hatchery-origin fish from offspring generation, but keep NAs (so many NAs due to GOD event)
```{r}
invisible(sapply(grep(pattern = "18", x = sillyvec, value = TRUE), function(silly) {
  my.gcl = get(paste0(silly, ".gcl"))
  
  # keep natural and NA to drop hatchery, don't != "YES"
  assign(
    x = paste0(silly, ".gcl"),
    value = filter(my.gcl, `Otolith Mark Present` == "NO" | is.na(`Otolith Mark Present`)),
    pos = 1
  )
  
}))
```

### Verify removed fish

Verify that we did as intended, check sillys for otolith read counts
```{r}
sapply(sillyvec, function(silly) {  # sillyvec_parentage
  my.gcl = get(paste0(silly, ".gcl"))
  my.gcl %>%
    select(`Otolith Mark Present`) %>% 
    mutate(silly = silly)
}, simplify = FALSE) %>%
  dplyr::bind_rows() %>%
  count(silly, `Otolith Mark Present`) %>%
  pivot_wider(names_from = `Otolith Mark Present`,
              values_from = n,
              values_fill =  0)
```

Save, even though we didn't split sillys
```{r}
if(!dir.exists("../../Genotypes/All_Streams_16_18_develop/OceanAK_Origin")) {
  dir.create("../../Genotypes/All_Streams_16_18_develop/OceanAK_Origin")
}

save_sillys(sillyvec = sillyvec, path = "../../Genotypes/All_Streams_16_18_develop/OceanAK_Origin")
```

# QA

Go through our standard Quality Assurance (QA) process, explore heterozygosity, and produce a matrix with sample sizes.
```{r qa_setup}
(
  sample_size_qa <- tibble(silly = sillyvec) %>%  # sillyvec_parentage
    mutate(genotyped = silly_n.GCL(sillyvec = sillyvec) %>% pull(n))  # sillyvec_parentage
)
```

## Filter missing

Remove individuals missing >=20% of genotypes (i.e. the 80% rule)
```{r qa_missing}
miss_loci <-
  RemoveIndMissLoci.GCL(sillyvec = sillyvec,  # sillyvec_parentage
                        proportion = 0.8,
                        loci = loci298)

save_objects(objects = "miss_loci",
             path =  "../../Objects/All_Streams_16_18_develop/")

# show individuals removed
# miss_loci$IDs_Removed

(
  sample_size_qa <- sample_size_qa %>%
    mutate(
      missing = genotyped - silly_n.GCL(sillyvec = sillyvec) %>% pull(n)  # sillyvec_parentage
    )
)
```

Of the 5,565 individuals removed, almost half were from PSTOCK16, we lost ~25% of those samples!

## Filter duplicates

Remove duplicate individuals within the same collection. Typically, we specify *duplicates* as a pair of individuals that share >=95% of genotypes. Once a pair of *duplicates* is identified, we keep the individual with the most genotypes and remove the other(s).

Identify duplicates
```{r qa_duplicate}
duplicate_check_95 <-
  CheckDupWithinSilly.GCL(
    sillyvec = sillyvec,
    loci = loci298,
    quantile = NULL,
    minproportion = 0.95,
    minnonmissing = 0.6,
    ncores = 4
  )

save_objects("duplicate_check_95", path = "../../Objects/All_Streams_16_18_develop/")
```

Remove **both** duplicates! As opposed to GSI work, where we want to keep individuals but aren't typically worried about paired data, here we want to remove both individuals as the paired data integrity is lost.
```{r qa_remove_duplicates}
RemoveDups.GCL(dupcheck = duplicate_check_95, remove_both = TRUE)

(
  sample_size_qa <- sample_size_qa %>%
    mutate(
      duplicate = genotyped - missing - silly_n.GCL(sillyvec = sillyvec) %>% pull(n)
    )
)
```

## Filter for heterozygosity

We know from the QC work that the **GTscore** genotyping pipeline can sometimes result in individuals with an overabundance of heterozygous calls (perhaps from contamination). We want to remove individuals with untrustworthy genotypes. To explore the utility of a *heterozygosity filter*, we should first plot a histogram of heterozygosity per individual to see what the problem looks like. From there we can set a threshold or range of acceptable heterozygosities.

Create individual heterozygosity function

**NOTE** this funciton assumes that you are using original sillys (relies on SillySource), not split by origin
```{r}
individual_heterozygosity <- function(sillyvec, loci = LocusControl$locusnames, ncores = 4) {
  
  start.time <- Sys.time() 
  
  if(!all(loci %in% LocusControl$locusnames)){
    
    stop(paste0("'", setdiff(loci, LocusControl$locusnames), "' from argument 'loci' not found in 'LocusControl' object!!!"))
    
  }
  
  if(!require("pacman")) install.packages("pacman"); library(pacman); pacman::p_load(tidyverse, doParallel, parallel, foreach)  # Install packages, if not in library and then load them
  
  if(ncores > parallel::detectCores()) {
    
    stop("'ncores' is greater than the number of cores available on machine\nUse 'detectCores()' to determine the number of cores on your machine")
    
  }
  
  all.gcl <- sapply(sillyvec, function(silly) {get(paste0(silly, ".gcl"), pos = 1)}, simplify = FALSE)
  
  cl <- parallel::makePSOCKcluster(ncores)
  
  doParallel::registerDoParallel(cl, cores = ncores)
  
  # Start parallel loop
  het <- foreach::foreach(silly = sillyvec, .packages = c("tidyverse")) %dopar% {
    
    my.gcl <- all.gcl[[silly]]
    
    dose1 <- my.gcl %>%
      dplyr::select(SillySource, dplyr::all_of(loci)) %>%
      tidyr::pivot_longer(
        cols = -SillySource,
        names_to = "locus",
        values_to = "dose1",
        values_drop_na = TRUE  # this drops any no-calls, which are NAs
      )
    
    dose2 <- my.gcl %>%
      dplyr::select(SillySource, dplyr::all_of(paste0(loci, ".1"))) %>%
      dplyr::rename_at(vars(paste0(loci, ".1")), ~ loci) %>%
      tidyr::pivot_longer(
        cols = -SillySource,
        names_to = "locus",
        values_to = "dose2",
        values_drop_na = TRUE  # this drops any no-calls, which are NAs
      )
    
    if(all.equal(dose1$SillySource, dose2$SillySource) & all.equal(dose1$locus, dose2$locus)) {
      
      my.gcl_tall <- dplyr::bind_cols(dose1, select(dose2, dose2))  # faster
      
    } else {
      
      my.gcl_tall <- dplyr::left_join(x = dose1, y = dose2, by = c("SillySource", "locus"))  # safer
      
    }
    
    my.gcl_tall %>%
      dplyr::group_by(SillySource) %>%
      dplyr::summarise(het = sum(dose1 != dose2) / length(dose1))
    
  } %>%
    dplyr::bind_rows()
  
  parallel::stopCluster(cl)  # end parallel loop
  
  output <- het %>%
    tidyr::separate(
      col = SillySource,
      into = c("silly", "fish_id"),
      sep = "_",
      remove = FALSE
    ) %>% 
    dplyr::mutate(fish_id = as.numeric(fish_id)) %>% 
    dplyr::arrange(silly, fish_id)
  
  print(Sys.time() - start.time)
  
  return(output)
}
```

Calculate heterozygosity

```{r heterozygosity_calculations}
(
  het_tibble <- individual_heterozygosity(sillyvec = sillyvec, loci = loci298, ncores = 4)
)

save_objects(objects = "het_tibble", path = "../../Objects/All_Streams_16_18_develop/")
```

### Plot heterozygosities

Plot heterozygosities by lineage with *tidyverse* for ease of use/visualization. Cutoffs (red vertical lines) are based on the +/- 1.5 IQR method that it is well known and well documented in literature.

#### Overall

Compare distribution of heterozygosities calculated by the *new* method vs. the *old*, naive method
**NOTE** no need to incorporate lineage here, since all collections here are even-year
```{r}
# NOTE: Original plotting by silly removed and +/- 1.5 IQR heterozygosity cutoff instilled -- After discussions among Chris, Tyler, Dan, and Kyle, they decided it would be more appropriate to examine heterozygosity by lineage (even and/or odd) instead of year. It was also agreed that the +/- 3SD cutoff (Anderson et al. 2010) was not restrictive enough as the distribution has a long right-sided tail and the upper bound for the cutoff should be closer to the main distribution. A hybrid right-side mirror approach based on the +/- 3SD was also suggested by Chris and discarded in favor of the +/- 1.5 IQR.

het_cutoffs <- het_tibble %>% 
  summarise(
    het_mean = mean(het),
    het_1.5IQR_low = quantile(het, probs = 0.25) - 1.5 * diff(quantile(het, probs = c(0.25, 0.75))),
    het_1.5IQR_high = quantile(het, probs = 0.75) + 1.5 * diff(quantile(het, probs = c(0.25, 0.75)))
  ) %>% 
  mutate(lineage = "even")

het_tibble %>%
  ggplot2::ggplot(aes(x = het)) +
  ggplot2::geom_histogram(binwidth = 1 / length(loci)) +
  ggplot2::facet_grid(rows = vars(lineage)) +
  ggplot2::geom_vline(data = het_cutoffs, aes(xintercept = het_mean, colour = "mean")) +
  ggplot2::geom_vline(data = het_cutoffs, aes(xintercept = het_1.5IQR_low, colour = "IQR")) +
  ggplot2::geom_vline(data = het_cutoffs, aes(xintercept = het_1.5IQR_high, colour = "IQR")) +
  ggplot2::scale_color_manual(name = "statistics", values = c(mean = "blue", IQR = "red")) +
  ggplot2::xlim(0, 1) +
  ggplot2::xlab("Individual Heterozygosity") +
  ggplot2::ylab("Frequency") +
  ggplot2::ggtitle("Individual Heterozygosity by Collection") +
  ggplot2::theme_bw() 
```

#### Compare silly

See how the *new* method would apply across original sillys
```{r fig.height=10}
# het_silly_means <- het_tibble %>%
#   group_by(og_silly) %>%
#   summarise(
#     het_mean = mean(het),
#     het_1.5IQR_low = quantile(het, probs = 0.25) - 1.5 * diff(quantile(het, probs = c(0.25, 0.75))),
#     het_1.5IQR_high = quantile(het, probs = 0.75) + 1.5 * diff(quantile(het, probs = c(0.25, 0.75)))
#   )

het_tibble %>%
  ggplot2::ggplot(aes(x = het)) +
  ggplot2::geom_histogram(binwidth = 1 / length(loci)) +
  ggplot2::facet_grid(rows = vars(silly), scales = "free_y") +
  ggplot2::geom_vline(data = het_cutoffs, aes(xintercept = het_mean, colour = "mean")) +
  ggplot2::geom_vline(data = het_cutoffs, aes(xintercept = het_1.5IQR_low, colour = "IQR")) +
  ggplot2::geom_vline(data = het_cutoffs, aes(xintercept = het_1.5IQR_high, colour = "IQR")) +
  ggplot2::scale_color_manual(name = "statistics", values = c(mean = "blue", IQR = "red")) +
  ggplot2::xlim(0, 1) +
  ggplot2::xlab("Individual Heterozygosity") +
  ggplot2::ylab("Frequency") +
  ggplot2::ggtitle("Individual Heterozygosity by Collection") +
  ggplot2::theme_bw(base_size = 12) 
```

### Remove fish outside of the IQR cutoffs

Count the number of fish to remove from each silly
```{r table_het_cutoff}
# Calculate 1.5 IQR
het_cutoffs

# Get list of fish to remove
het_tibble %>%
  dplyr::mutate(lineage = "even") %>% 
  dplyr::left_join(het_cutoffs, by = "lineage") %>%
  dplyr::filter(het < het_1.5IQR_low | het > het_1.5IQR_high) %>%
  dplyr::count(silly)
```

Filter for fish to remove and save
```{r remove_het_cutoff}
(
  hets_removed <- het_tibble %>%
    dplyr::mutate(lineage = "even") %>%
    dplyr::left_join(het_cutoffs, by = "lineage") %>%
    dplyr::filter(het < het_1.5IQR_low | het > het_1.5IQR_high)
)

save_objects("hets_removed", "../../Objects/All_Streams_16_18_develop/")
```

Remove these fish
**NOTE** `silly` and `fish_id` come from `SillySource`, this only works if you are using the original sillys (i.e. you haven't split them by origin or something).
```{r}
sapply(unique(hets_removed$silly), function(mysilly) {
  
  het_ids <- filter(hets_removed, silly == mysilly) %>% pull(fish_id)
  
  RemoveIDs.GCL(silly = mysilly, IDs = het_ids)
})

(
  sample_size_qa <- sample_size_qa %>%
    mutate(
      heterozygosity = genotyped - missing - duplicate - silly_n.GCL(sillyvec = sillyvec) %>% pull(n)
    )
)
```

## Final dataset

Count how many fish we retained
```{r qa_final}
(sample_size_qa <- sample_size_qa %>%
   mutate(final =  silly_n.GCL(sillyvec = sillyvec) %>% pull(n)))

save_objects("sample_size_qa", path = "../../Objects/All_Streams_16_18_develop/")
```

Save the post-QA genotypes
```{r save_postQA_genotypes}
if(!dir.exists("../../Genotypes/All_Streams_16_18_develop/OceanAK_Origin_PostQA")) {
  dir.create("../../Genotypes/All_Streams_16_18_develop/OceanAK_Origin_PostQA")
}

save_sillys(sillyvec, "../../Genotypes/All_Streams_16_18_develop/OceanAK_Origin_PostQA/")
```

# Export genepop file

Now that we have final post QA/QC genotypes, we want to save our work by exporting the genotypes as a **Genepeop** file using `gcl2Genepop.GCL`, with individuals named by *SillySource*.
```{r genepop}
if(!dir.exists("../../Genepop/All_Streams_16_18_develop")) {
  dir.create("../../Genepop/All_Streams_16_18_develop")
}

gcl2Genepop.GCL(sillyvec = sillyvec, 
                loci = loci298, 
                path = "../../Genepop/All_Streams_16_18_develop/All_Streams_16_18_develop_postQA.gen", 
                VialNums = TRUE, 
                usat = FALSE,
                ncores = 4)
```

# Convert genepop to FRANz format

The **Genepop** file is fed into Chase's *genepop2franz.GCL.r* script to create a **FRANz** input file for parentage analysis.
```{r Convert genepop to Franz for parentage}
# detach(package:reshape2, unload = TRUE)
source("../genepop2franz.GCL.r")
genepop2franz.GCL(
  Genepop = "../../Genepop/All_Streams_16_18_develop/All_Streams_16_18_develop_postQA.gen",
  OceanAK = "../../OceanAK/AHRP Salmon Biological Data 20220511_174609_names.csv",
  Year = c(2016, 2018),
  Stream = c("HOGAN", "STOCK", "GILMOUR", "PADDY", "ERB"),
  output_dir = "../"
)
```

**NOTE** Since we want to run all 5 streams together, you need to go in and edit the *FRANz* input file to reflect 1 sampling location and take out delimiters between locations.

Final sample size
```{r}
sum(sample_size_qa$final)
```

End
