==> Stockdale_Hogan_Gilmour_Paddy_Erb/2_Parentage_All_Streams_2016_2018_develop.Rmd <==
---
title: "All Streams 2016/2018 Parentage"
subtitle: "Hogan + Stockdale + Gilmour + Paddy + Erb"
author: "Kyle Shedd"
date: "26 May 2022"
output:
  html_notebook:
    theme: united
    toc: yes
editor_options: 
  chunk_output_type: inline
---

```{r setup, include=FALSE}
rm(list = ls())
library(coin)
library(MASS)
library(scales)
library(tidyverse)
library(leaflet)
library(lubridate)
library(ggExtra)

knitr::opts_chunk$set(out.width = "100%", fig.width = 10)
```

# **Background**

**NOTE** the Gilmour, Erb, and Paddy data will need to be re-run post-GOD. All of the `NA`-origin parents (2016) were removed in `1_QA`, and thus not included in these analyses. The `NA`-origin offspring (2018) were retained.

## Pedigree reconstruction

We use the program *FRANz* [link](https://www.bioinf.uni-leipzig.de/Software/FRANz/About.html) to reconstruct parent-offspring relationships. Below are analyses of parentage for **all fitness streams (Hogan, Stockdale, Gilmour, Paddy, and Erb)** for the **even lineage** years **2016** (potential parents) and **2018** (potential offspring).

**NOTE** due to memory allocation limits on the UAF cluster, we had to break this into two (2) separate analyses:
  1) only Hogan 2016 parents (n = 11,522), all streams 2018 offspring  (n = 25,926), and 
  2) no Hogan 2016 parents (just Stockdale, Gilmour, Paddy, and Erb; n = 12,973), all streams 2018 offspring (n = 25,926)

Ideally, offspring will only have parentage assignments from a single stream...

*FRANz* was run by Chase secretly on the UAF cluster on 5/23/2022 using the following parameters:
FRANz.exe --Nmmax 25000 --Nfmax 25000 --femrepro 1:2 --malerepro 1:2 --typingerror 0.006 --updatefreqs --poutformat 2 All_Streams_16_18_develop_postQA_2016_2018_Hogan_parents_only.dat

*FRANz* was run by Chase secretly on the UAF cluster on 5/24/2022 using the following parameters:
FRANz.exe --Nmmax 25000 --Nfmax 25000 --femrepro 1:2 --malerepro 1:2 --typingerror 0.006 --updatefreqs --poutformat 2 All_Streams_16_18_develop_postQA_2016_2018_no_Hogan_parents.dat

The parameters are defined as follows:

* --Nmmax and --Nfmax are the maximum numbers of candidate mothers and fathers. To obtain our values, we used an estimated escapement of ~50,000 and divided by 2. This escapement number came from very generously rounding up the maximum population estimate (aerial, stream walk, stream walk corrected for distance surveyed) from each stream. These estimates are listed under in the Excel file found here: [link](V:\Documents\5_Coastwide\Multispecies\AHRP\Field data\Sample Summary Extract and Genotype.xlsx).    
* --femrepro and --malerepro specify the age range in which an individual can reproduce  
* --typingerror refers to the overall genotyping error rate, where ours is usually ~0.5% (can obtained from QC; 0.006 has been used as a default of sorts)
* --updatefreqs specifies that *FRANz* should update allele frequencies using MCMC sampling  
* --poutformat specifies that all potential parents should be listed, not just the most likely  
* --fullsibtest tests for full siblings among offspring   

We do not use --fullsibtest; it won't save output (Kyle's Genetics Domain Computer) or it maxes memory (Kristen's computer).

All output files can also be found here: Output is saved here: [link](V:\Analysis\5_Coastwide\Multispecies\Alaska Hatchery Research Program\PWS Pink\Franz\All_Streams_16_18_develop). The summary file provides information about the power of our marker suite. According to the FRANz manual, marker sets are not considered powerful if these cumulative exclusion probabilities are less than 0.95, indicating the probability  a random pair of individuals in the population has a 5% chance of having a genotype pair compatible to an offspring genotype. Since nearly all of our probabilities are 1, we can be confident in the power of our 298 amplicons to make parent-offspring assignments.  

## Data import and overview

  * Read in the following data files:  
    1) "parentage.csv" file for parentage assignments produced by *FRANz*  
    2) paired genotype and OceanAK data (otolith reads, some field data), and  
    3) paired StreamSpecimens contractor data (gill condition, preyed upon, etc.) and `riverdist` data (stream distance)

Since we had to do two *FRANz* runs for memory allocation reasons, we need to combine the `parentage.csv` outputs before we proceed.
```{r}
parentage_16_18_hogan_parents <-
  readr::read_csv(
    "~/../Desktop/Local_PWS_pinks/Franz/All_Streams_16_18_develop/2016_2018_HoganParentsOnly/results/parentage.csv"
  ) %>%  # read_csv("../../Franz/All_Streams_16_18_develop/2016_2018_HoganParentsOnly/results/parentage.csv") 
  dplyr::mutate(run = "only_hogan")

parentage_16_18_no_hogan_parents <-
  readr::read_csv(
    "~/../Desktop/Local_PWS_pinks/Franz/All_Streams_16_18_develop/2016_2018_NoHoganParents/results/parentage.csv"
  ) %>%  # read_csv("../../Franz/All_Streams_16_18_develop/2016_2018_NoHoganParents/results/parentage.csv")
  dplyr::mutate(run = "no_hogan")

(
  parentage_16_18 <-
    dplyr::bind_rows(
      parentage_16_18_hogan_parents,
      parentage_16_18_no_hogan_parents
    )
)

base::rm(parentage_16_18_hogan_parents, parentage_16_18_no_hogan_parents)
```

Import paired *OceanAK* data and `riverdist` stream location data.
```{r data_import, echo=FALSE, warning=FALSE, message=FALSE}
paired_16_18 <-
  readr::read_csv(
    "../Franz/All_Streams_16_18_develop_postQA_OceanAK_paired_2016_2018_HOGAN_STOCK_GILMOUR_PADDY_ERB.csv"
  )
stream_specimens_riverdist <-
  readr::read_csv(
    "~/../Desktop/Local_PWS_pinks/GIS/stream_specimens_riverdist_all_streams_2013_2020.csv"
  )  # read_csv("../GIS/R/all_streams/stream_specimens_riverdist_all_streams_2013_2020.csv")
```

  * Plot a histogram of *FRANz* parentage posterior probabilities to show strength of assignments
  
```{r FRANz_posterior_plot, echo=FALSE, warning=FALSE, message=FALSE}
# first, we establish colorblindness friendly palettes
cbPalette1 <- c("#0072b2", "#d55e00", "#F0E442", "#CC79A7") # blue, orange, yellow, and pink
cbPalette2 <- c("#CC79A7", "#0072b2") # pink and blue for natural- and hatchery-origin

# plot the posteriors
parentage_16_18  %>% 
  dplyr::filter(!base::is.na(`Parent 1`)) %>% 
  ggplot2::ggplot(aes(x = Posterior)) +
  ggplot2::geom_histogram(breaks = seq(0, 1, 0.01), colour = "black") +
  ggplot2::theme_bw() +
  ggplot2::labs(x = "Posterior Probability",
                y = "Count") +
  ggplot2::ggtitle("Histogram of FRANz posterior probabilities for parental assignments") +
  ggplot2::facet_grid(rows = dplyr::vars(run))
```

**Results:** Almost all parental assignments have a posterior probability of 1, which is very robust.  

## Subsetting paired genetic/field data

### Parentage

Filter parentage data from *FRANz* to only keep assignments with a posterior probability > 0.9, make tidy/tall (one row per parent-offspring assignment).
```{r filter_parentage_data, echo=FALSE, warning=FALSE, message=FALSE}
# Filter for posterior > 0.9 + make tidy relative to assignment (1 row per parentage assignment)
(
  parentage_16_18_filter <- parentage_16_18 %>%
    dplyr::filter(Posterior > 0.9) %>%
    dplyr::select(Offspring, `Parent 1`, `Parent 2`, run) %>%
    tidyr::pivot_longer(
      names_to = "Parent",
      values_to = "Parent_ID",
      cols = -c(Offspring, run)
    ) %>%
    dplyr::filter(!is.na(Parent_ID))
)
```

Since we are combining runs ("only Hogan parents" and "no Hogan parents"), we need to verify that Offspring were only assigned parents in 1 of those two runs (i.e. not both!). Going to check two different ways to make extra, super-duper sure :).
```{r}
tmp_only_hogan <- parentage_16_18_filter %>% 
  dplyr::filter(run == "only_hogan") %>% 
  dplyr::pull(Offspring)

tmp_no_hogan <- parentage_16_18_filter %>% 
  dplyr::filter(run == "no_hogan") %>% 
  dplyr::pull(Offspring)

base::intersect(tmp_only_hogan, tmp_no_hogan)

base::rm(tmp_no_hogan, tmp_only_hogan)

parentage_16_18_filter %>% 
  dplyr::filter(Parent == "Parent 1") %>% 
  dplyr::count(Offspring, run) %>% 
  dplyr::count(Offspring) %>% 
  dplyr::count(n, name = "count")
```

### OceanAK

The file containing paired genotype and OceanAK data has a lot of information we do not need at this time (e.g. genotypes for each marker). We pull out only the columns that contain identifying information for each individual.

First view all the non-genotype columns to see what we want to retain.
```{r view_oceanak_paired_data, echo=FALSE, warning=FALSE, message=FALSE}
# Find the non-genotype columns
paired_16_18 %>% 
  dplyr::select(-tidyselect::starts_with("Ogo_RAD")) %>% 
  utils::head()
```

Remove genotype columns, rename, and reformat others.
```{r subset_oceanak_paired_data, echo=FALSE, warning=FALSE, message=FALSE}
# Filter for non-genotype columns
(
  paired_16_18_filter <- paired_16_18 %>%
    dplyr::select(
      franz_id,
      SILLY,
      `Fish ID`,
      `DNA Tray Code`,
      `DNA Tray Well Code`,
      `Location Code`,
      `Sample Year`,
      `Sample Date`,
      SEX,
      `Length Mm`,
      `Otolith Mark Present`,
      `Otolith Mark ID`
    ) %>%
    dplyr::rename(
      silly = SILLY,
      fish_id = "Fish ID",
      dna_tray_code = "DNA Tray Code",
      dna_tray_well_code = "DNA Tray Well Code",
      year = "Sample Year",
      length = "Length Mm",
      otolith_mark_present = "Otolith Mark Present",
      otolith_mark_id = "Otolith Mark ID"
    ) %>%
    dplyr::mutate(
      stream = stringr::str_remove(string = `Location Code`, pattern = " Creek"),
      date = lubridate::ymd(`Sample Date`),
      DOY = lubridate::yday(date),  # day of year
      sex = dplyr::case_when(SEX == "M" ~ "Male",
                             SEX == "F" ~ "Female"),
      origin = dplyr::case_when(
        otolith_mark_present == "NO" ~ "Natural",
        otolith_mark_present == "YES" ~ "Hatchery"
      )  # add origin variable
    ) %>%
    dplyr::mutate(origin = base::factor(origin, c(
      "Natural", "Hatchery"
    ))) %>%  # make factor to ensure hatchery != red
    tidyr::unite(
      col = "sample",
      c("dna_tray_code", "dna_tray_well_code"),
      sep = "_",
      remove = FALSE
    ) %>%
    dplyr::select(
      franz_id,
      stream,
      year,
      origin,
      sex,
      date,
      DOY,
      length,
      otolith_mark_present,
      otolith_mark_id,
      silly,
      fish_id,
      sample,
      dna_tray_code,
      dna_tray_well_code
    )
)
```

### Stream Specimens + `riverdist`

The file containing paired genotype and OceanAK data has a lot of information we do not need at this time (e.g. genotypes for each marker). We pull out only the columns that contain identifying information for each individual.

First view all the non-genotype columns to see what we want to retain.
```{r view_stream_specimens_riverdist_data, echo=FALSE, warning=FALSE, message=FALSE}
stream_specimens_riverdist %>% 
  utils::head()
```

Join with *OceanAK* paired data, rename, and reformat columns.
```{r subset_stream_specimens_riverdist_paired_data, echo=FALSE, warning=FALSE, message=FALSE}
(
  paired_16_18_filter <- paired_16_18_filter %>%
    dplyr::left_join(stream_specimens_riverdist, by = c("sample" = "Sample")) %>%  # join based on DWP_well
    dplyr::select(-ADFGStreamCode) %>%
    dplyr::rename(
      stream_trib = StreamName,
      spawning_state = SpawningState,
      pre_spawn = PreSpawn,
      partial_spawn = PartialSpawner,
      preyed_upon = PreyedUpon,
      latitude = Latitude,
      longitude = Longitude,
      distance_mouth = mouthdist,
      distance_tide = dist2tide,
      high_tide = hitide
    ) %>%
    dplyr::mutate(
      intertidal = dplyr::case_when(
        distance_tide <= 0 ~ "Intertidal",
        distance_tide > 0 ~ "Upstream",
        is.na(distance_tide) ~  NA_character_
      )  # create intertidal
    ) %>%
    dplyr::select(
      franz_id,
      stream,
      year,
      origin,
      sex,
      date,
      DOY,
      length,
      distance_mouth,
      intertidal,
      distance_tide,
      tidyselect::everything()
    )
)
```

Save paired data.
```{r save_paired_filter}
readr::write_csv(paired_16_18_filter, "all_streams_paired_16_18_filter.csv")
```

Clean up environment to save memory.
```{r}
base::rm(paired_16_18, stream_specimens_riverdist, parentage_16_18)
```

## Sample sizes

  * Quick look at sample sizes for potential parents (2016)

**NOTE** Gilmour, Erb, and Paddy were all affected by GOD. We *dropped* the `NA` origin parents (2016) in `1_QA`
```{r sample_sizes_parents, echo=FALSE, warning=FALSE, message=FALSE}
paired_16_18_filter %>% 
  dplyr::filter(year == 2016) %>%  # filter for parent year
  dplyr::count(stream, origin) %>% 
  tidyr::pivot_wider(names_from = origin, values_from = n, values_fill = 0)
```

  * Quick look at sample sizes for potential offspring (2018)

**NOTE** Gilmour, Erb, and Paddy were all affected by GOD. We *kept* the `NA` origin offspring (2018) in `1_QA`
```{r sample_sizes_offspring, echo=FALSE, warning=FALSE, message=FALSE}
paired_16_18_filter %>% 
  dplyr::filter(year == 2018) %>%  # filter for offspring year
  dplyr::count(stream)
```

Which ones were involved in GOD?
```{r}
paired_16_18_filter %>% 
  dplyr::filter(year == 2018) %>%  # filter for offspring year
  dplyr::count(stream, origin) %>% 
  tidyr::pivot_wider(names_from = origin, values_from = n, values_fill = 0)
```

  * Sample size plots of potential parents and potential offspring collected by day for each stream

**NOTE** Since we already dropped the `NA` parents (2016) for Gilmour, Erb, and Paddy, they do not show up as grey columns here in these plots.
```{r sample_sizes_plot, echo=FALSE, warning=FALSE, message=FALSE}
sample_size_plot <-
  function(indata = paired_XX_XX_filter, title = "Stream Year-Year") {
    base::sapply(base::unique(indata$stream), function(x) {
      indata %>%
        dplyr::filter(stream == x) %>%
        ggplot2::ggplot(aes(x = DOY, fill = origin)) +
        ggplot2::geom_bar(colour = "black") +  # group = interaction(DOY, origin))
        ggplot2::scale_x_continuous(breaks = seq(180, 320, by = 5),
                           labels = format(x = (
                             as.Date("2013-12-31") + seq(180, 320, by = 5)
                           ), "%b %d")) +
        ggplot2::facet_grid(year ~ stream) +
        ggplot2::geom_hline(yintercept = 0, colour = "black") +
        ggplot2::labs(title = title, 
                      fill = "Origin",
                      x = "Sample Date",
                      y = "Number of Samples") +
        ggplot2::theme_bw() +
        ggplot2::theme(text = element_text(size = 15),
              axis.text.x = element_text(angle = 90)) +
        ggplot2::scale_fill_manual(values = cbPalette2)
    }, simplify = FALSE)
  }

sample_size_plot(indata = paired_16_18_filter, title = "All Streams 2016-2018")
```

**NOTE** `NA` origin fish in 2018 (offspring) for Erb, Gilmour, and Paddy are due to the GOD event.

# **Basic Statistics**

## Joining parentage with subsetted paired data

Here, we joined parentage data from *FRANz* with individual paired data so that we can match data to each offspring and parent. For this dataset **one row per parent-offspring assignment**. 

**Note** that any column ending in `_off` refers to offspring and `_par` refers to parent. 
```{r join_parentage_and_paired_objects, echo=FALSE, warning=FALSE, message=FALSE}
# Creating a single, "tidy" object where each row is a parent-offspring relationship
# Offspring with 2 parents have 2 rows
# _off data is offspring, _par data is parents
(
  parents_paired_16_18 <- parentage_16_18_filter %>%
    dplyr::left_join(paired_16_18_filter, by = c("Offspring" = "franz_id")) %>%
    dplyr::left_join(
      paired_16_18_filter,
      by = c("Parent_ID" = "franz_id"),
      suffix = c("_off", "_par")
    ) %>% 
    dplyr::rename(offspring_id = Offspring,
                  parent = Parent,
                  parent_id = Parent_ID)
)

readr::write_csv(x = parents_paired_16_18,
                 file = "all_streams_parents_paired_16_18.csv")
```

  * Number of single-parent offspring (dyad) and two-parent offspring (triad) assignments for each stream

```{r n_dyads_triads}
parents_paired_16_18 %>% 
  dplyr::count(stream_par, parent) %>% 
  tidyr::pivot_wider(names_from = parent, values_from = n)
```

  * Percent of offspring sampled that were assigned parents (from any stream) for each stream

```{r}
n_off <- paired_16_18_filter %>%
  dplyr::filter(year == 2018) %>%  # filter for offspring year
  dplyr::count(stream)

n_off_assigned <- parents_paired_16_18 %>%
  dplyr::filter(parent == "Parent 1") %>% 
  dplyr::count(stream_off)  # accounts for offspring straying among streams

dplyr::left_join(n_off,
                 n_off_assigned,
                 by = c("stream" = "stream_off"),
                 suffix = c("_samp", "_assign")) %>%
  dplyr::mutate(percent_off_assign = base::round(n_assign / n_samp * 100, 1)) %>% 
  dplyr::select(stream, percent_off_assign)

rm(n_off, n_off_assigned)
```

  * Number of offspring (from any stream) assigned by parent origin for each stream

```{r n_offspring_by_origin}
parents_paired_16_18 %>% 
  dplyr::count(stream_par, origin_par) %>% 
  tidyr::pivot_wider(names_from = origin_par, values_from = n)
```

  * Number of distinct, successful (assigned offspring) parents by origin for each stream

```{r}
parents_paired_16_18 %>% 
  dplyr::distinct(parent_id, stream_par, origin_par) %>% 
  dplyr::count(stream_par, origin_par) %>% 
  tidyr::pivot_wider(names_from = origin_par, values_from = n)
```

  * Number of distinct, successful (assigned offspring) hatchery-origin parents by hatchery code for each stream

```{r}
parents_paired_16_18 %>% 
  dplyr::filter(origin_par == "Hatchery") %>% 
  dplyr::distinct(parent_id, stream_par, otolith_mark_id_par) %>% 
  dplyr::count(stream_par, otolith_mark_id_par) %>% 
  tidyr::pivot_wider(names_from = otolith_mark_id_par, values_from = n)
```

**Results:** Most offspring were only assigned to 1 parent (dyads), but there were a decent amount assigned to 2 parents (triads). Offspring assignment rates ranged from a low of **12.2%** in Paddy to a high of **34.8%** in Hogan. Interestingly, Stockdale and Gilmour (on Montague Island) mostly receive strays from AFK, Hogan receives a mix of AFK (most common), CCH, WNH, and SGH, and Paddy and Erb receive most of their strays from WNH, followed by AFK, CCH, and a few SGH.

## Proportion test

We performed a chi-square test on the data to naively determine whether the proportions of offspring assigned to hatchery- and natural-origin parents significantly differ from the proportions of hatchery- and natural-origin parents sampled. 

**Note** that for this analysis, the data were not divided out by sex, so offspring from triads are double counted. 
```{r chi_square_test, echo=FALSE, warning=FALSE, message=FALSE}
# How many potential spawners from each origin?
spawners_16 <- paired_16_18_filter %>%
  dplyr::filter(year == 2016) %>%  # filter for parent year
  dplyr::count(stream, origin) %>%
  dplyr::mutate(generation = "Parents")

# How many offspring from each origin?
offspring_18 <- parents_paired_16_18 %>%
  dplyr::count(stream_par, origin_par) %>%
  dplyr::rename(origin = origin_par,
                stream = stream_par) %>%
  dplyr::mutate(generation = "Offspring")

# Make tidy and spread
(
  spawners_and_offspring_16_18 <-
    dplyr::bind_rows(spawners_16, offspring_18) %>%
    dplyr::arrange(stream, origin) %>%
    tidyr::pivot_wider(names_from = generation, values_from = n)
)

rm(spawners_16, offspring_18)  # keep it clean!

# Run chisq.test by stream
spawners_and_offspring_16_18 %>%
  dplyr::nest_by(stream) %>%
  dplyr::mutate(model = list(stats::chisq.test(x = data[,-1]))) %>%
  dplyr::summarise(
    chisq = model$statistic,
    df = model$parameter,
    p_value = model$p.value,
    method = model$method,
    .groups = "keep"
  ) %>%
  dplyr::mutate(signif = dplyr::case_when(p_value <= 0.05 ~ TRUE,
                                          p_value > 0.05 ~ FALSE)) %>%
  dplyr::select(stream, chisq, df, p_value, signif, method)
```

**Results:** All streams, except for Gilmour and Paddy, showed a significant difference in the proportions of offspring assigned to hatchery- and natural-origin parents relative to the proportions of potential parents sampled, indicating an under-representation of offspring assigning to hatchery-origin parents. 

# **Reproductive Success (RS)**

## Single parent-offspring pairs (dyads)

Here, we calculate family size for parents of both sexes. We also plot two sets of histograms of the proportion of parents with a given family size by sex (number of offspring assigned to a female parent or a male parent) and origin. Next, we calculate sex- and origin-specific mean reproductive success (RS; table below and vertical lines in second set of histograms) and relative reproductive success (RRS) with 95% CIs. For both RS and RRS metrics, we include sampled fish from the parental generation to which offspring were not assigned in addition to known parents (zeroes).

### RS distribution for successful parents (no zeroes)

This first plot does **not** include zeroes, of which there will be many, but it provides better resolution among successful parents, especially for visualizing max family size.
```{r family_size_by_sex_without_zeros, echo=FALSE, warning=FALSE, message=FALSE, fig.height=10}
# Plot histogram of proportion of successful parents with a given family size
parents_paired_16_18 %>% 
  dplyr::count(stream_par, parent_id, origin_par, sex_par) %>% 
  dplyr::count(stream_par, sex_par, origin_par, n, name = "n_par") %>% 
  dplyr::group_by(stream_par, sex_par, origin_par) %>% 
  dplyr::mutate(p = n_par / sum(n_par)) %>% 
  ggplot2::ggplot(aes(x = n, y = p, fill = origin_par)) +
  ggplot2::geom_col(position = position_dodge2(preserve="single"), colour = "black") +
  ggplot2::scale_y_continuous(breaks = scales::pretty_breaks()) +
  ggplot2::scale_x_continuous(breaks = scales::pretty_breaks()) +
  ggplot2::facet_grid(stream_par ~ sex_par) +
  ggplot2::labs(title="Distribution of Family Size by Successful Parents by Sex",
                fill = "Parent: Origin",
                x = "Number of Offspring",
                y = "Proportion of Parents") +
  ggplot2::scale_fill_manual(values = cbPalette2) +
  ggplot2::theme_bw()
```

### RS for all parents (including zeros)

Here, we transform the data to **one row per parent** with `n` = number of offspring assigned, and include all potential parents (including those assigned zero offspring). 
```{r paired_filter_parents, echo=FALSE, warning=FALSE, message=FALSE}
# Transform to one row per parent with n = number of offspring produced
(
  paired_16_18_filter_parents <- paired_16_18_filter %>%
    dplyr::filter(year == "2016") %>%
    dplyr::left_join(
      x = .,
      y = dplyr::count(parents_paired_16_18, parent_id),
      by = c("franz_id" = "parent_id")
    ) %>% 
    dplyr::mutate(n = tidyr::replace_na(data = n, replace = 0)) %>% 
    dplyr::select(franz_id, n, tidyselect::everything())
)

write_csv(x = paired_16_18_filter_parents, file = "all_streams_paired_16_18_filter_parents.csv")
```

  * RS including sampled fish from the parental generation to which offspring were not assigned.

```{r calculate_RS, echo=FALSE, warning=FALSE, message=FALSE}
# Mean RS, including zeroes
rs_16_18_0 <- paired_16_18_filter_parents %>%
  dplyr::group_by(stream, origin, sex) %>% 
  dplyr::summarise(RS = mean(n, na.rm = TRUE), .groups = "drop")

rs_16_18_0 %>%
  tidyr::pivot_wider(names_from = origin, values_from = RS, names_prefix = "RS_") %>% 
  dplyr::mutate(RRS = RS_Hatchery / RS_Natural) %>% 
  dplyr::select(stream, sex, RS_Hatchery, RS_Natural, RRS) %>% 
  dplyr::mutate(dplyr::across(where(is.numeric), round, 2))
```

### RS distribution for all parents (including zeroes)

This plot **does** include zeroes.
```{r family_size_by_sex_with_zeros, echo=FALSE, warning=FALSE, message=FALSE, fig.height=10}
# Plot histogram of proportion of all parents with a given family size
paired_16_18_filter_parents %>% 
  dplyr::count(stream, sex, origin, n, name = "n_par") %>% 
  dplyr::group_by(stream, sex, origin) %>% 
  dplyr::mutate(p = n_par / sum(n_par)) %>% 
  ggplot2::ggplot(aes(x = n, y = p, fill = origin)) +
  ggplot2::geom_col(position = position_dodge2(preserve="single"), colour = "black") +
  ggplot2::scale_y_continuous(breaks = scales::pretty_breaks()) +
  ggplot2::scale_x_continuous(breaks = scales::pretty_breaks()) +
  ggplot2::facet_grid(stream ~ sex) +
  ggplot2::labs(title="Distribution of Family Size by All Parents by Sex",
                fill = "Parent: Origin",
                x = "Number of Offspring",
                y = "Proportion of Parents") +
  ggplot2::scale_fill_manual(values = cbPalette2) +
  ggplot2::theme_bw()
```

**Results:** RS distributions are highly skewed, similar to a negative binomial. In the histograms, note the large number of potential parents assigned no offspring (RS = 0) that were absent from the family size plots above, which only considered successful parents (RS > 0).

### Relative reproductive success (RRS)

#### *Calculate RRS*

Caclulcate RRS by stream and sex.
```{r calculate RRS}
rs_16_18_0 %>%
  tidyr::pivot_wider(names_from = origin, values_from = RS, names_prefix = "RS_") %>% 
  dplyr::mutate(RRS = RS_Hatchery / RS_Natural) %>% 
  dplyr::select(stream, sex, RRS) %>% 
  dplyr::mutate(dplyr::across(where(is.numeric), round, 2)) %>% 
  tidyr::pivot_wider(names_from = sex, values_from = RRS)
```

**Results:** RRS estimates are < 1 for all streams, although some estimates are likely not significantly < 1.

#### *Test for significant differences in RRS*

Next, we test for significant differences in RS between hatchery- and natural-origin fish. This is equivalent to testing if RRS < 1. We perform two different statistical tests on the results in the top panel, including

  1. a non-parametric permutation test (aka randomization test) and
  2. a parametric negative binomial GLM (with no additional covariates).

P-values are presented for each statistical test.

##### By sex

We'll test for difference separately for each sex (dyad data).
```{r RRS_statistical_testing_by_sex}
paired_16_18_filter_parents %>%
  dplyr::nest_by(stream, sex) %>%
  dplyr::mutate(permutation = list(
    coin::oneway_test(
      n ~ origin,
      data = data,
      distribution = coin::approximate(nresample = 10000),
      alternative = "two.sided"  # for a one-sided test use "greater"
    )
  ),
  glm_nb = list(MASS::glm.nb(
    n ~ origin,
    data = data,
    init.theta = 1,
    link = log
  ))) %>%
  dplyr::summarise(permutation_p_value = as.numeric(coin::pvalue(permutation)),
                   glm_nb_p_value = summary(glm_nb)$coefficients["originHatchery", "Pr(>|z|)"],
                   .groups = "drop") %>% 
  dplyr::mutate(dplyr::across(where(is.numeric), round, 4))
```

**Results:** There are *only* significant differences in RRS for females in Erb, and *only* in Erb, Hogan, and Paddy for males. 

##### Both sexes combined

...and again with both sexes together (offspring in triads are double counted).
```{r RRS_statistical_testing_by_stream}
paired_16_18_filter_parents %>%
  dplyr::nest_by(stream) %>%
  dplyr::mutate(permutation = list(
    coin::oneway_test(
      n ~ origin,
      data = data,
      distribution = coin::approximate(nresample = 10000),
      alternative = "two.sided"  # for a one-sided test use "greater"
    )
  ),
  glm_nb = list(MASS::glm.nb(
    n ~ origin,
    data = data,
    init.theta = 1,
    link = log
  ))) %>%
  dplyr::summarise(permutation_p_value = as.numeric(coin::pvalue(permutation)),
                   glm_nb_p_value = summary(glm_nb)$coefficients["originHatchery", "Pr(>|z|)"],
                   .groups = "drop") %>% 
  dplyr::mutate(dplyr::across(where(is.numeric), round, 4))
```

**Results:** There are significant differences in RRS for Erb, Hogan, and Stockdale^*^ when we ignore sex-specific differences. 

#### *RRS plus 95% CIs for females and males*

We calculated 95% CIs for RRS estimated above, based on [Kalinowski and Taper (2005)](http://www.nrcresearchpress.com/doi/10.1139/f04-239), a maximum likelihood estimate based on [Hinrichsen (2003)](https://cdnsciencepub.com/doi/10.1139/f03-070). This method has been used in other peer reviewed papers, such as [Ford et al. (2016)](https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0164801).

Create function.
```{r RRS_confidence_intervals_function, echo=FALSE, warning=FALSE, message=FALSE}
rrs_ci_kalinowski <- function(n_h_off, n_w_off, n_h_par, n_w_par, alpha = .05) {
    
    chi_alpha <- stats::qchisq(p = (1 - alpha), df = 1)
    n_off <- sum(c(n_h_off, n_w_off))
    n_par <- sum(c(n_h_par, n_w_par))
    
    rs_h <- n_h_off / n_h_par
    rs_w <- n_w_off / n_w_par
    
    p_h_par <- n_h_par / n_par
    p_w_par <- n_w_par / n_par
    
    rrs_h <- rs_h / rs_w
    rrs_w <- rs_w / rs_w
    rrs_avg <- (rrs_h * p_h_par) + (rrs_w * p_w_par)
    
    rrs_ml <-
      (n_h_off * log(p_h_par * rrs_h / rrs_avg)) + (n_w_off * log(p_w_par * rrs_w / rrs_avg))
    
    xi_dist <-
      bind_rows(lapply(seq(0.01, 5, by = 0.01), function(rrs_h_xi) {
        rrs_avg_xi <- (rrs_h_xi * p_h_par) + (rrs_w * p_w_par)
        tibble(
          rrs_crit = rrs_h_xi,
          logl = (n_h_off * log(p_h_par * rrs_h_xi / rrs_avg_xi)) + (n_w_off * log(p_w_par * rrs_w / rrs_avg_xi)) - (rrs_ml - chi_alpha / 2)
        )
      }))
    
    rrs_min <- xi_dist %>%
      mutate(abs_logl = abs(logl)) %>%
      filter(rrs_crit < rrs_h) %>%
      top_n(-1, abs_logl) %>%
      pull(rrs_crit)
    
    rrs_max <- xi_dist %>%
      mutate(abs_logl = abs(logl)) %>%
      filter(rrs_crit > rrs_h) %>%
      top_n(-1, abs_logl) %>%
      pull(rrs_crit)
    
    # xi_plot <- xi_dist %>%
    #   ggplot(aes(x = rrs_crit, y = logl)) +
    #   geom_line() +
    #   geom_hline(yintercept = 0, colour = "red", lwd = 2) +
    #   geom_vline(xintercept = c(rrs_h, rrs_min, rrs_max), colour = "blue") +
    #   ylim(c(-5, 5)) +
    #   xlim(c(0, 2)) +
    #   ylab("Log Likelihood - Chi Sq Value") +
    #   annotate("text", x = rrs_h + 0.1, y = xi_dist %>% filter(rrs_crit == xi_dist$rrs_crit[which.min(abs(xi_dist$rrs_crit  - rrs_h))]) %>% pull(logl) + 0.4, label = round(rrs_h, 2)) +
    #   annotate("text", x = rrs_min - 0.1, y = xi_dist %>% filter(rrs_crit == rrs_min) %>% pull(logl) + 0.4, label = rrs_min) +
    #   annotate("text", x = rrs_max + 0.1, y = xi_dist %>% filter(rrs_crit == rrs_max) %>% pull(logl) + 0.4, label = rrs_max)
    #
    # print(xi_plot)
    return(c(rrs_min, rrs_h, rrs_max))
  }
```

Calculate RRS with 95% CI's
```{r RRS_confidence_intervals}
paired_16_18_filter_parents %>%
  dplyr::rename(n_off = n) %>%
  dplyr::group_by(stream, sex, origin) %>%
  dplyr::summarise(n_par = dplyr::n(),
                   n_off = sum(n_off),
                   .groups = "drop") %>%
  tidyr::pivot_wider(names_from = origin, values_from = c(n_par, n_off)) %>%
  dplyr::rename(
    n_h_off = n_off_Hatchery,
    n_w_off = n_off_Natural,
    n_h_par = n_par_Hatchery,
    n_w_par = n_par_Natural
  ) %>%
  dplyr::nest_by(stream, sex) %>%
  dplyr::mutate(rrs_cis = list(
    rrs_ci_kalinowski(
      n_h_off = data$n_h_off,
      n_w_off = data$n_w_off,
      n_h_par = data$n_h_par,
      n_w_par = data$n_w_par
    )
  )) %>%
  dplyr::summarise(
    rrs_lower = rrs_cis[1],
    rrs = rrs_cis[2],
    rrs_upper = rrs_cis[3],
    .groups = "drop"
  ) %>%
  dplyr::mutate(dplyr::across(where(is.numeric), round, 2))
```

**Results:** RRS estimates are significant (upper 95% CI does not contain 1) for all streams and sexes except for Gilmour females and Paddy males.

### RS by field data

Now that we've calculated RRS, we can look into the relationship between RS and some of our field data (date, length, location, etc.).

#### *RS by sample date*

View how RS varies by parent sample date across the season.

**NOTE** the y-axis varies among streams to facilitate visualizing patterns.
```{r RS_by_sample_date, echo=FALSE, warning=FALSE, message=FALSE, fig.height=10}
# Females and males separated
paired_16_18_filter_parents %>%
  ggplot2::ggplot(aes(x = date, y = n, fill = origin)) +
  ggplot2::geom_hline(yintercept = 0, color = "grey50") +
  ggplot2::geom_jitter(
    height = 0,
    size = 4,
    shape = 21,
    color = "black",
    alpha = 0.5
  ) +
  ggplot2::geom_smooth(
    method =  "loess",
    formula = 'y ~ x',
    color = "black",
    size = 1,
    alpha = 0.7
  ) +
  ggplot2::facet_grid(stream ~ sex, scales = "free_y") +
  ggplot2::labs(title = "Reproductive Success by Parental Sample Date and Sex",
                fill = "Parent: Origin",
                x = "Parental Sample Date",
                y = "Reproductive Success") +
  ggplot2::theme_bw() +
  # ggplot2::theme(panel.grid.minor = element_blank()) +
  ggplot2::scale_fill_manual(values = cbPalette2)
```

**Results:** Trends vary by stream, but there is not a ton of separation in RS between hatchery- and natural-origin fish regardless of parent sample date.

#### *RS by body length*

View how RS varies by parent body length (MEH).

**NOTE** the y-axis varies among streams to facilitate visualizing patterns.
```{r RS_by_body_length, echo=FALSE, warning=FALSE, message=FALSE, fig.height=10}
# Females and males separated
paired_16_18_filter_parents %>%
  dplyr::filter(length > 300) %>%  # remove outlier from Stockdale female
  ggplot2::ggplot(aes(x = length, y = n, fill = origin)) +
  ggplot2::geom_hline(yintercept = 0, color = "grey50") +
  ggplot2::geom_jitter(
    height = 0,
    size = 4,
    shape = 21,
    color = "black",
    alpha = 0.5
  ) +
  ggplot2::geom_smooth(
    method =  "loess",
    formula = 'y ~ x',
    color = "black",
    size = 1,
    alpha = 0.7
  ) +
  ggplot2::facet_grid(stream ~ sex, scales = "free_y") +
  ggplot2::labs(title = "Reproductive Success by Parental Body Length and Sex",
                fill = "Parent: Origin",
                x = "Parental Body Length (mm)",
                y = "Reproductive Success") +
  ggplot2::theme_bw() +
  # ggplot2::theme(panel.grid.minor = element_blank()) +
  ggplot2::scale_fill_manual(values = cbPalette2)
```

**Results:** Trends vary by stream, but in general, larger fish tend to have higher RS.

#### *RS by stream location*

View how RS varies by parent sample location within the stream (distance from stream mouth in m).

**NOTE** both the x-axis and y-axis vary among streams to facilitate visualizing patterns.

**NOTE** this is a messy plot, will need to figure out how to display this better some other time...tricky because each stream is fairly different in length and some have multiple tributaries (hence the multiple high-tide lines).
```{r RS_by_location, echo=FALSE, warning=FALSE, message=FALSE, fig.height=10}
# Females and males separated
RS_by_location_plots <-
  sapply(unique(paired_16_18_filter_parents$stream), function(x)
  {
    paired_16_18_filter_parents %>%
      filter(stream == x) %>%
      ggplot2::ggplot(aes(x = distance_mouth, y = n, fill = origin)) +
      ggplot2::geom_hline(yintercept = 0, color = "grey50") +
      ggplot2::geom_jitter(
        height = 0,
        size = 4,
        shape = 21,
        color = "black",
        alpha = 0.5
      ) +
      ggplot2::geom_smooth(
        method =  "loess",
        formula = 'y ~ x',
        color = "black",
        size = 1,
        alpha = 0.7
      ) +
      ggplot2::geom_vline(aes(xintercept = high_tide)) +
      ggplot2::facet_grid(stream ~ sex, scales = "free_y") +
      ggplot2::labs(title = "Reproductive Success by Parental Location and Sex",
                    fill = "Parent: Origin",
                    x = "Parental Sample Distance from Stream Mouth (m)",
                    y = "Reproductive Success") +
      ggplot2::theme_bw() +
      ggplot2::theme(title = element_blank(),
                     axis.title.x = element_blank(),
                     axis.title.y = element_blank()) +
      # ggplot2::theme(panel.grid.minor = element_blank()) +
      ggplot2::scale_fill_manual(values = cbPalette2)
  }, simplify = FALSE)



gridExtra::grid.arrange(RS_by_location_plots$Erb,
                        RS_by_location_plots$Gilmour,
                        RS_by_location_plots$Hogan,
                        RS_by_location_plots$Paddy,
                        RS_by_location_plots$Stockdale,
                        nrow = 5,
                        top = "Reproductive Success by Parental Location and Sex",
                        left = "Reproductive Success",
                        bottom = "Parental Sample Distance from Stream Mouth (m)")
```

**Results:** Trends vary by stream, but in general, RS is higher near the intertidal for Erb, Paddy, and Stockdale. Not so much for Hogan and Gilmour.

#### *RS by intertidal*

View how RS varies by parent sample location within the stream (intertidal or upstream).

**NOTE** the y-axis varies among streams to facilitate visualizing patterns, labels indicate mean values.
```{r RS_by_intertidal, echo=FALSE, warning=FALSE, message=FALSE, fig.height=10}
# Females and males separated
paired_16_18_filter_parents %>%
  dplyr::filter(!is.na(intertidal)) %>% 
  ggplot2::ggplot(aes(x = intertidal, y = n, fill = origin)) +
  ggplot2::geom_hline(yintercept = 0, color = "grey50") +
  ggplot2::geom_boxplot() +
  ggplot2::stat_summary(
    ggplot2::aes(label = round(..y.., 2)),
    fun.y = "mean",
    position = ggplot2::position_dodge(width = 1.2),
    geom = "text",
    vjust = -1
  ) +
  ggplot2::facet_grid(stream ~ sex, scales = "free_y") +
  ggplot2::labs(title = "Reproductive Success by Parental Location (Intertidal) and Sex",
                fill = "Parent: Origin",
                x = "Stream Location: Intertidal vs. Upstream",
                y = "Reproductive Success") +
  ggplot2::theme_bw() +
  # ggplot2::theme(panel.grid.minor = element_blank()) +
  ggplot2::scale_fill_manual(values = cbPalette2)
```

**Results:** Trends vary by stream, but in general, RS is higher near the intertidal. Regardless of intertidal/upstream location, RS tends to be higher for natural-origin fish relative to hatchery-origin fish.

#### *RS by river segment*

View how RS varies by parent sample location within the stream (river segment).

**NOTE:** Hogan and Erb only have a single `riverdist` segment, so they are not considered here.

##### Gilmour

```{r RS_by_riverdist_seg_gilmour, echo=FALSE, warning=FALSE, message=FALSE, fig.height=5}
# Females and males separated
paired_16_18_filter_parents %>%
  dplyr::filter(stream == "Gilmour",
                !is.na(riverdist_seg)) %>% 
  dplyr::mutate(river_seg = factor(riverdist_seg)) %>% 
  dplyr::count(river_seg, origin, name = "n_par") %>% 
  tidyr::pivot_wider(names_from = origin, values_from = n_par)


paired_16_18_filter_parents %>%
  dplyr::filter(stream == "Gilmour",
                !is.na(riverdist_seg)) %>% 
  dplyr::mutate(river_seg = factor(riverdist_seg)) %>% 
  ggplot2::ggplot(aes(x = river_seg, y = n, fill = origin)) +
  ggplot2::geom_hline(yintercept = 0, color = "grey50") +
  ggplot2::geom_boxplot() +
  ggplot2::stat_summary(
    ggplot2::aes(label = round(..y.., 2)),
    fun.y = "mean",
    position = ggplot2::position_dodge(width = 1.2),
    geom = "text",
    vjust = -1
  ) +
  ggplot2::facet_grid(stream ~ sex, scales = "free_y") +
  ggplot2::labs(title = "Reproductive Success by River Segment and Sex: Gilmour",
                fill = "Parent: Origin",
                x = "Riverdist Segment",
                y = "Reproductive Success") +
  ggplot2::theme_bw() +
  # ggplot2::theme(panel.grid.minor = element_blank()) +
  ggplot2::scale_fill_manual(values = cbPalette2)
```

##### Stockdale

```{r RS_by_riverdist_seg_stockdale, echo=FALSE, warning=FALSE, message=FALSE, fig.height=5}
# Females and males separated
paired_16_18_filter_parents %>%
  dplyr::filter(stream == "Stockdale",
                !is.na(riverdist_seg)) %>% 
  dplyr::mutate(river_seg = factor(riverdist_seg)) %>% 
  dplyr::count(river_seg, origin, name = "n_par") %>% 
  tidyr::pivot_wider(names_from = origin, values_from = n_par)


paired_16_18_filter_parents %>%
  dplyr::filter(stream == "Stockdale",
                !is.na(riverdist_seg)) %>% 
  dplyr::mutate(river_seg = factor(riverdist_seg)) %>% 
  ggplot2::ggplot(aes(x = river_seg, y = n, fill = origin)) +
  ggplot2::geom_hline(yintercept = 0, color = "grey50") +
  ggplot2::geom_boxplot() +
  ggplot2::stat_summary(
    ggplot2::aes(label = round(..y.., 2)),
    fun.y = "mean",
    position = ggplot2::position_dodge(width = 1.2),
    geom = "text",
    vjust = -1
  ) +
  ggplot2::facet_grid(stream ~ sex, scales = "free_y") +
  ggplot2::labs(title = "Reproductive Success by River Segment and Sex: Stockdale",
                fill = "Parent: Origin",
                x = "Riverdist Segment",
                y = "Reproductive Success") +
  ggplot2::theme_bw() +
  # ggplot2::theme(panel.grid.minor = element_blank()) +
  ggplot2::scale_fill_manual(values = cbPalette2)
```

##### Paddy

```{r RS_by_riverdist_seg_paddy, echo=FALSE, warning=FALSE, message=FALSE, fig.height=5}
# Females and males separated
paired_16_18_filter_parents %>%
  dplyr::filter(stream == "Paddy",
                !is.na(riverdist_seg)) %>% 
  dplyr::mutate(river_seg = factor(riverdist_seg)) %>% 
  dplyr::count(river_seg, origin, name = "n_par") %>% 
  tidyr::pivot_wider(names_from = origin, values_from = n_par)


paired_16_18_filter_parents %>%
  dplyr::filter(stream == "Paddy",
                !is.na(riverdist_seg)) %>% 
  dplyr::mutate(river_seg = factor(riverdist_seg)) %>% 
  ggplot2::ggplot(aes(x = river_seg, y = n, fill = origin)) +
  ggplot2::geom_hline(yintercept = 0, color = "grey50") +
  ggplot2::geom_boxplot() +
  ggplot2::stat_summary(
    ggplot2::aes(label = round(..y.., 2)),
    fun.y = "mean",
    position = ggplot2::position_dodge(width = 1.2),
    geom = "text",
    vjust = -0.2
  ) +
  ggplot2::facet_grid(stream ~ sex, scales = "free_y") +
  ggplot2::labs(title = "Reproductive Success by River Segment and Sex: Paddy",
                fill = "Parent: Origin",
                x = "Riverdist Segment",
                y = "Reproductive Success") +
  ggplot2::theme_bw() +
  # ggplot2::theme(panel.grid.minor = element_blank()) +
  ggplot2::scale_fill_manual(values = cbPalette2)
```

#### *RS by pre-spawn*

View how RS varies by pre-spawning state within the stream (TRUE or FALSE).

**NOTE** the y-axis varies among streams to facilitate visualizing patterns, labels indicate mean values.
```{r RS_by_prespawn, echo=FALSE, warning=FALSE, message=FALSE, fig.height=10}
# Females and males separated
paired_16_18_filter_parents %>%
  dplyr::filter(!is.na(pre_spawn)) %>%
  ggplot2::ggplot(ggplot2::aes(x = pre_spawn, y = n, fill = origin)) +
  ggplot2::geom_hline(yintercept = 0, color = "grey50") +
  ggplot2::geom_boxplot() +
  ggplot2::stat_summary(
    ggplot2::aes(label = round(..y.., 2)),
    fun.y = "mean",
    position = ggplot2::position_dodge(width = 1.2),
    geom = "text",
    vjust = -1
  ) +
  ggplot2::facet_grid(stream ~ sex, scales = "free_y") +
  ggplot2::labs(title = "Reproductive Success by Pre-spawn Status and Sex",
                fill = "Parent: Origin",
                x = "Pre-spawn Mortality?",
                y = "Reproductive Success") +
  ggplot2::theme_bw() +
  # ggplot2::theme(panel.grid.minor = element_blank()) +
  ggplot2::scale_fill_manual(values = cbPalette2)
```

**Results:** Trends vary by stream, but in general, RS is higher for fish that were not marked as pre-spawn mortality.

#### *RS by partial-spawn*

View how RS varies by partial-spawning state within the stream (TRUE or FALSE).

**NOTE** the y-axis varies among streams to facilitate visualizing patterns, labels indicate mean values.
```{r RS_by_partialspawn, echo=FALSE, warning=FALSE, message=FALSE, fig.height=10}
# Females and males separated
paired_16_18_filter_parents %>%
  dplyr::filter(!is.na(partial_spawn)) %>%
  ggplot2::ggplot(ggplot2::aes(x = partial_spawn, y = n, fill = origin)) +
  ggplot2::geom_hline(yintercept = 0, color = "grey50") +
  ggplot2::geom_boxplot() +
  ggplot2::stat_summary(
    ggplot2::aes(label = round(..y.., 2)),
    fun.y = "mean",
    position = ggplot2::position_dodge(width = 1.2),
    geom = "text",
    vjust = -1
  ) +
  ggplot2::facet_grid(stream ~ sex, scales = "free_y") +
  ggplot2::labs(title = "Reproductive Success by Partial-spawn Status and Sex",
                fill = "Parent: Origin",
                x = "Partial-spawn Mortality?",
                y = "Reproductive Success") +
  ggplot2::theme_bw() +
  # ggplot2::theme(panel.grid.minor = element_blank()) +
  ggplot2::scale_fill_manual(values = cbPalette2)
```

**Results:** Trends vary by stream, but in general, RS is higher for fish that were not marked as partial-spawn mortality.

#### *RS by preyed-upon*

View how RS varies by partial-spawning state within the stream (TRUE or FALSE).

**NOTE** the y-axis varies among streams to facilitate visualizing patterns, labels indicate mean values.
```{r RS_by_preyed_upon, echo=FALSE, warning=FALSE, message=FALSE, fig.height=10}
# Females and males separated
paired_16_18_filter_parents %>%
  dplyr::filter(!is.na(preyed_upon)) %>%
  ggplot2::ggplot(ggplot2::aes(x = preyed_upon, y = n, fill = origin)) +
  ggplot2::geom_hline(yintercept = 0, color = "grey50") +
  ggplot2::geom_boxplot() +
  ggplot2::stat_summary(
    ggplot2::aes(label = round(..y.., 2)),
    fun.y = "mean",
    position = ggplot2::position_dodge(width = 1.2),
    geom = "text",
    vjust = -1
  ) +
  ggplot2::facet_grid(stream ~ sex, scales = "free_y") +
  ggplot2::labs(title = "Reproductive Success by Preyed-upon Status and Sex",
                fill = "Parent: Origin",
                x = "Preyed-upon Mortality?",
                y = "Reproductive Success") +
  ggplot2::theme_bw() +
  # ggplot2::theme(panel.grid.minor = element_blank()) +
  ggplot2::scale_fill_manual(values = cbPalette2)
```

**Results:** Trends vary by stream, but in general, RS is higher for fish that were not marked as preyed-upon mortality.

## Parent pair-offspring trios (triads)

Because we have numerous individuals assigned two parents, we examined instances of four potential types of crosses. Crosses may be between hatchery fish (HH), between natural fish (NN), and between natural and hatchery fish, with NH indicating a natural-origin mother and hatchery-origin father and HN indicating a hatchery-origin mother and natural-origin father. 

### Cross types observed

Need to reformat data so that we have **one row per offspring**, with both `sire` and `dam` data preserved.
```{r cross_as_new_grouping_variable, echo=FALSE, warning=FALSE, message=FALSE}
# Which offspring have 2 parents?
triads_16_18 <- parents_paired_16_18 %>% 
  dplyr::filter(parent == "Parent 2") %>% 
  dplyr::pull(offspring_id)

(
  parents_paired_16_18_cross <- parents_paired_16_18 %>%
    dplyr::filter(offspring_id %in% triads_16_18) %>%
    dplyr::select(-parent) %>% # drop "Parent 1" and "Parent 2", replace with dam/sire
    dplyr::mutate(
      relationship = dplyr::case_when(sex_par == "Male" ~ "sire",
                                      sex_par == "Female" ~ "dam")
    ) %>%
    tidyr::pivot_wider(
      names_from = relationship,
      values_from = c(parent_id, tidyselect::contains("_par"))
    ) %>%  # pivot based on dam/sire
    dplyr::rename_with(~ base::gsub(
      x = .x,
      pattern = "_par",
      replacement = ""
    )) %>%
    tidyr::unite(
      col = "mating_id",
      c("parent_id_dam", "parent_id_sire"),
      sep = "_",
      remove = FALSE
    ) %>%
    dplyr::mutate(
      cross_type = dplyr::case_when(
        origin_dam == "Natural" & origin_sire == "Natural" ~ "NN",
        origin_dam == "Natural" & origin_sire == "Hatchery" ~ "NH",
        origin_dam == "Hatchery" & origin_sire == "Natural" ~ "HN",
        origin_dam == "Hatchery" & origin_sire == "Hatchery" ~ "HH"
      )
    )
)

rm(triads_16_18)

readr::write_csv(x = parents_paired_16_18_cross, file = "all_streams_parents_paired_16_18_cross.csv")
```

Verify that parents were sampled in the same stream.
```{r parents_from_same_stream}
all(parents_paired_16_18_cross$stream_dam == parents_paired_16_18_cross$stream_sire)
```

### RS for cross types

Great, now need to reformat data so that we have **one row per unique mating**, with both `sire` and `dam` data preserved, count number of offspring.
```{r paired_16_18_filter_cross}
# Join to get 1 row per unique mating with dam/sire data
(
  paired_16_18_filter_cross <- parents_paired_16_18_cross %>%
    dplyr::select(-c(
      offspring_id, tidyselect::contains("_off")
    )) %>%
    dplyr::distinct() %>%  # drop to one row per mating
    dplyr::left_join(x = .,
                     y = dplyr::count(parents_paired_16_18_cross, mating_id),
                     by = "mating_id") %>%  # join in n_offspring
    dplyr::rename(stream = stream_dam,
                  year = year_dam) %>%
    dplyr::select(-c(stream_sire, year_sire)) %>%
    dplyr::select(
      mating_id,
      n,
      cross_type,
      stream,
      year,
      tidyselect::everything()
    )
)

readr::write_csv(x = paired_16_18_filter_cross, file = "all_streams_paired_16_18_filter_cross.csv")
```

**Results:** We have **617** unique matings documented among the five fitness streams for 2016.

Calculate mean RS by stream and cross type
```{r calculate_RS_cross, echo=FALSE, warning=FALSE, message=FALSE}
# Mean RS, including zeroes
rs_16_18_cross <- paired_16_18_filter_cross %>%
  dplyr::group_by(stream, cross_type) %>% 
  dplyr::summarise(RS = mean(n, na.rm = TRUE), .groups = "drop")

rs_16_18_cross %>%
  tidyr::pivot_wider(names_from = cross_type,
                     values_from = RS,
                     names_prefix = "RS_") %>%
  dplyr::mutate(
    RRS_HH = RS_HH / RS_NN,
    RRS_HN = RS_HN / RS_NN,
    RRS_NH = RS_NH / RS_NN
  ) %>%
  dplyr::mutate(dplyr::across(where(is.numeric), round, 2))
```

**Results:** RRS values by cross type are a bit all over the place, but looking at the sample sizes below, that isn't terribly surprising given the pretty sparse sample sizes by cross type.

#### Samples sizes

How many unique matings occurred by cross type and stream?
```{r n_matings by cross}
paired_16_18_filter_cross %>%
  dplyr::count(stream, cross_type, name = "n_mating") %>%
  tidyr::pivot_wider(names_from = cross_type, values_from =  n_mating)
```

**Results:** Sample sizes of families (matings) by cross type are fairly small, especially for non-NN crosses.

How many offspring where assigned to each cross type and stream?
```{r n_offspring by cross}
paired_16_18_filter_cross %>% 
  dplyr::group_by(stream, cross_type) %>% 
  dplyr::summarise(n = sum(n), .groups = "drop") %>% 
  tidyr::pivot_wider(names_from = cross_type, values_from =  n)
```

### RS distribution for all crosses (no zeroes)

This plot does **not** include zeroes, because we can not infer a cross occurred unless offspring are assigned.
```{r family_size_by_cross, echo=FALSE, warning=FALSE, message=FALSE, fig.height=10}
# Plot histogram of proportion of all parents with a given family size
paired_16_18_filter_cross %>% 
  dplyr::count(stream, cross_type, n, name = "n_mating") %>% 
  dplyr::group_by(stream, cross_type) %>% 
  dplyr::mutate(p = n_mating / sum(n_mating)) %>% 
  ggplot2::ggplot(aes(x = n, y = p, fill = cross_type)) +
  ggplot2::geom_col(position = position_dodge2(preserve="single"), colour = "black") +
  ggplot2::scale_y_continuous(breaks = scales::pretty_breaks()) +
  ggplot2::scale_x_continuous(breaks = scales::pretty_breaks(n = 10)) +
  ggplot2::facet_grid(stream ~ .) +
  ggplot2::labs(title="Distribution of Family Size by Cross Type",
                fill = "Parent: Cross Type",
                x = "Number of Offspring",
                y = "Proportion of Families") +
  ggplot2::scale_fill_manual(values = cbPalette1) +
  ggplot2::theme_bw()
```

### Relative reproductive success (RRS)

#### *Calculate RRS*

Caclulcate RRS by stream and cross type, all relative to NN.
```{r calculate RRS by cross type}
rs_16_18_cross %>%
  tidyr::pivot_wider(names_from = cross_type,
                     values_from = RS,
                     names_prefix = "RS_") %>%
  dplyr::mutate(
    RRS_HH = RS_HH / RS_NN,
    RRS_HN = RS_HN / RS_NN,
    RRS_NH = RS_NH / RS_NN
  ) %>%
  dplyr::select(stream, contains("RRS")) %>%
  dplyr::rename_with( ~ base::gsub(
    x = .x,
    pattern = "RRS_",
    replacement = ""
  )) %>%
  dplyr::mutate(dplyr::across(where(is.numeric), round, 2))
```

**Results:** RRS estimates are not necessarily < 1 for all streams, although sample sizes for non-NN cross types are pretty low.

#### *RRS plus 95% CIs for cross types*

##### HH/NN

Calculate RRS with 95% CI's
```{r RRS_confidence_intervals_by_cross_type_HH}
paired_16_18_filter_cross %>%
  dplyr::filter(cross_type %in% c("HH", "NN")) %>% 
  dplyr::rename(n_off = n) %>%
  dplyr::group_by(stream, cross_type) %>%
  dplyr::summarise(n_fam = dplyr::n(),
                   n_off = sum(n_off),
                   .groups = "drop") %>%
  tidyr::pivot_wider(names_from = cross_type, values_from = c(n_fam, n_off)) %>%
  dplyr::rename(
    n_h_off = n_off_HH,
    n_w_off = n_off_NN,
    n_h_par = n_fam_HH,
    n_w_par = n_fam_NN
  ) %>%
  dplyr::nest_by(stream) %>%
  dplyr::mutate(rrs_cis = list(
    rrs_ci_kalinowski(
      n_h_off = data$n_h_off,
      n_w_off = data$n_w_off,
      n_h_par = data$n_h_par,
      n_w_par = data$n_w_par
    )
  )) %>%
  dplyr::summarise(
    rrs_lower = rrs_cis[1],
    rrs = rrs_cis[2],
    rrs_upper = rrs_cis[3],
    .groups = "drop"
  ) %>%
  dplyr::mutate(dplyr::across(where(is.numeric), round, 2))
```

**Results:** None of these RRS results are significant, low sample size of families + high variance in family size.

##### HN/NN

Calculate RRS with 95% CI's
```{r RRS_confidence_intervals_by_cross_type_HN}
paired_16_18_filter_cross %>%
  dplyr::filter(cross_type %in% c("HN", "NN")) %>% 
  dplyr::rename(n_off = n) %>%
  dplyr::group_by(stream, cross_type) %>%
  dplyr::summarise(n_fam = dplyr::n(),
                   n_off = sum(n_off),
                   .groups = "drop") %>%
  tidyr::pivot_wider(names_from = cross_type, values_from = c(n_fam, n_off)) %>%
  dplyr::rename(
    n_h_off = n_off_HN,
    n_w_off = n_off_NN,
    n_h_par = n_fam_HN,
    n_w_par = n_fam_NN
  ) %>%
  dplyr::nest_by(stream) %>%
  dplyr::mutate(rrs_cis = list(
    rrs_ci_kalinowski(
      n_h_off = data$n_h_off,
      n_w_off = data$n_w_off,
      n_h_par = data$n_h_par,
      n_w_par = data$n_w_par
    )
  )) %>%
  dplyr::summarise(
    rrs_lower = rrs_cis[1],
    rrs = rrs_cis[2],
    rrs_upper = rrs_cis[3],
    .groups = "drop"
  ) %>%
  dplyr::mutate(dplyr::across(where(is.numeric), round, 2))
```

**Results:** None of these RRS results are significant, low sample size of families + high variance in family size.

##### NH/NN

Calculate RRS with 95% CI's
```{r RRS_confidence_intervals_by_cross_type_NH}
paired_16_18_filter_cross %>%
  dplyr::filter(cross_type %in% c("NH", "NN")) %>% 
  dplyr::rename(n_off = n) %>%
  dplyr::group_by(stream, cross_type) %>%
  dplyr::summarise(n_fam = dplyr::n(),
                   n_off = sum(n_off),
                   .groups = "drop") %>%
  tidyr::pivot_wider(names_from = cross_type, values_from = c(n_fam, n_off)) %>%
  dplyr::rename(
    n_h_off = n_off_NH,
    n_w_off = n_off_NN,
    n_h_par = n_fam_NH,
    n_w_par = n_fam_NN
  ) %>%
  dplyr::nest_by(stream) %>%
  dplyr::mutate(rrs_cis = list(
    rrs_ci_kalinowski(
      n_h_off = data$n_h_off,
      n_w_off = data$n_w_off,
      n_h_par = data$n_h_par,
      n_w_par = data$n_w_par
    )
  )) %>%
  dplyr::summarise(
    rrs_lower = rrs_cis[1],
    rrs = rrs_cis[2],
    rrs_upper = rrs_cis[3],
    .groups = "drop"
  ) %>%
  dplyr::mutate(dplyr::across(where(is.numeric), round, 2))
```

**Results:** None of these RRS results are significant, low sample size of families + high variance in family size.

# **Mating dynamics**

We already verified that all of our matings took place between dams and sires that were sampled in the same streams, but how do the dams and sires differ in terms of sample date and sample location?

## Sire vs. Dam sample date

```{r dam_v_sire_sample_date, fig.height = 10}
paired_16_18_filter_cross %>% 
  ggplot2::ggplot(aes(x = date_dam, y = date_sire, fill = cross_type)) +
  ggplot2::geom_jitter(
    height = 0,
    size = 4,
    shape = 21,
    color = "black",
    alpha = 0.5
  ) +
  geom_abline(intercept = 0, slope = 1) +
  ggplot2::facet_grid(stream ~ ., scales = "fixed") +
  ggplot2::labs(title = "Distribution of Crosses by Dam and Sire Sample Date",
                fill = "Parent: Cross Type",
                x = "Dam Sample Date",
                y = "Sire Sample Date") +
  ggplot2::theme_bw() +
  ggplot2::theme(axis.text.x = element_text(angle = 90, vjust = 0.5)) +
  ggplot2::coord_fixed() +
  # ggplot2::theme(panel.grid.minor = element_blank()) +
  ggplot2::scale_fill_manual(values = cbPalette1)
```

**Results:** Samples dates for dams and sires align fairly well seem believable. In general, dams were sampled slightly later, which makes sense as males typically enter the system first.

## Sire vs. Dam sample location

```{r dam_v_sire_sample_location, fig.height = 10}
paired_16_18_filter_cross %>% 
  ggplot2::ggplot(aes(x = distance_mouth_dam, y = distance_mouth_sire, fill = cross_type)) +
  ggplot2::geom_jitter(
    height = 0,
    size = 4,
    shape = 21,
    color = "black",
    alpha = 0.5
  ) +
  geom_abline(intercept = 0, slope = 1) +
  ggplot2::facet_grid(stream ~ ., scales = "free_y") +
  ggplot2::labs(title = "Distribution of Crosses by Dam and Sire Sample Location",
                fill = "Parent: Cross Type",
                x = "Dam Sample Location",
                y = "Sire Sample Location") +
  ggplot2::theme_bw() +
  ggplot2::theme(axis.text.x = element_text(angle = 90, vjust = 0.5)) +
  # ggplot2::theme(panel.grid.minor = element_blank()) +
  ggplot2::scale_fill_manual(values = cbPalette1)
```

**Results:** This is messy, but it looks like sires tended to be sampled further downstream in Hogan and Stockdale.

## Multiple matings

We set out to discover who is mating with whom and when. We know multiple matings occur, but we want to understand whether it's males and/or females that have multiple partners. We thus set out to assess promiscuity within the sexes. We next calculated the number of mates for each known female and male parent (dams and sires).
```{r multiple_matings, echo=FALSE, warning=FALSE, message=FALSE}
# How many dams had multiple sires
dam_promiscuity <- parents_paired_16_18_cross %>% 
  dplyr::group_by(parent_id_dam) %>% 
  dplyr::summarise(n_sires = n_distinct(parent_id_sire)) %>% 
  dplyr::count(n_sires) %>% 
  dplyr::rename(n_mates = n_sires, n_dams = n)

# How many sires had multiple dams
sire_promiscuitiy <- parents_paired_16_18_cross %>% 
  dplyr::group_by(parent_id_sire) %>% 
  dplyr::summarise(n_dams = n_distinct(parent_id_dam)) %>% 
  dplyr::count(n_dams) %>% 
  dplyr::rename(n_mates = n_dams, n_sires = n)

(promiscuity <- dplyr::full_join(dam_promiscuity, sire_promiscuitiy, by = "n_mates"))

rm(dam_promiscuity, sire_promiscuitiy)
```
**Results:** Both dams and sires displayed promiscuity at relatively equal levels. 

# **Investigate natural straying**

How many offspring had parent from each stream?
```{r n_offspring_by_parent_stream}
parents_paired_16_18 %>%
  dplyr::filter(parent == "Parent 1") %>% 
  dplyr::mutate(
    stream_par = base::factor(
      x = stream_par,
      levels = c("Erb", "Paddy", "Hogan", "Gilmour", "Stockdale")
    ),
    stream_off = base::factor(
      x = stream_off,
      levels = c("Erb", "Paddy", "Hogan", "Gilmour", "Stockdale")
    )
  ) %>%
  dplyr::count(stream_par, stream_off) %>%  #, origin.par
  tidyr::pivot_wider(names_from = stream_off, values_from = n)
```

**Results:** Geographic patterns of straying among streams make sense. Paddy and Erb are mainland populations that are fairly close together, and Gilmour and Stockdale are very close together on Montague Island. Hogan definitely seems like an "attractor" stream, receiving strays from all five study streams (and relatively high pHOS).

Filter for parents that produced strays and determine how many of their offspring strayed.
```{r how_many_strays}
parents_produce_strays <- parents_paired_16_18 %>%
  dplyr::filter(parent == "Parent 1",
                stream_par != stream_off) %>%
  dplyr::distinct(parent_id) %>%
  dplyr::pull()

parents_paired_16_18 %>% 
  dplyr::filter(parent_id %in% parents_produce_strays) %>% 
  dplyr::group_by(parent_id, stream_par, origin_par) %>% 
  dplyr::summarise(n_home = sum(stream_par == stream_off),
                   n_stray = sum(stream_par != stream_off),
                   .groups = "keep") %>% 
  dplyr::mutate(n_off = sum(n_home, n_stray),
                p_stray = n_stray / n_off) 

parents_paired_16_18 %>% 
  dplyr::filter(parent_id %in% parents_produce_strays) %>% 
  dplyr::group_by(parent_id, stream_par, origin_par) %>% 
  dplyr::summarise(n_home = sum(stream_par == stream_off),
                   n_stray = sum(stream_par != stream_off),
                   .groups = "keep") %>% 
  dplyr::mutate(n_off = sum(n_home, n_stray),
                p_stray = n_stray / n_off) %>% 
  dplyr::group_by(stream_par, origin_par) %>% 
  dplyr::summarise(n_stray = sum(n_stray), .groups = "drop") %>% 
  tidyr::pivot_wider(names_from = origin_par, values_from = n_stray)


rm(parents_produce_strays)
```

Neat to see that sibling groups do not all tend to stray together, typically just 1 or 2 fish from a given parent end up straying. Keep in mind, that these aren't "stray rates" since sampling proportion varies across streams (and thus the probability of detection!).

Compare to how many total offspring assigned back to parents from each stream.
```{r n_offspring_by_origin_total}
parents_paired_16_18 %>% 
  dplyr::count(stream_par, origin_par) %>% 
  tidyr::pivot_wider(names_from = origin_par, values_from = n)
```

**Results:** No obvious influence of parent-origin on propensity of offspring to stray.

# **Summary of Results**

## 2016-2018 (Parentage)

1. Almost all exclusion probabilities were equal to 1, which means we can be confident in the ability of our marker set to correctly assign parents to offspring.  
2. 6,412 out of 25,926 offspring were assigned to 2,925 distinct parents, for an overall assignment rate of 28.6%.
4. RRS point estimates ranged from 0.30-1.01 for males and 0.45-1.31 for females. RRS estimates were statistically significant (95% CI did not contain 1) for all combinations, except Stockdale and Gilmour for males, and Gilmour and Paddy for females.  
5. All four crosses were represented (HH, NH, HN, and NN), but cross type RS relative to the NN were not significant, largely due to small sample sizes of families.  

```{r}
save.image("2_Parentage_All_Streams_2016_2018_develop_20220527.RData")
```

End
